Create a Microbiome Dataset

This is a constructor for the MbioDataset class. It creates a MbioDataset containing a list of Collections and a SampleMetadata object.

Usage

MbioDataset(collections, metadata, ontology)

# S4 method for missing,missing,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,SampleMetadata,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,SampleMetadata,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,SampleMetadata,character
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,data.frame,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,data.frame,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,data.frame,character
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,list,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,list,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,list,character
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,missing,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,missing,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,missing,character
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,character,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,character,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collections,character,character
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,SampleMetadata,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,SampleMetadata,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,SampleMetadata,character
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,data.frame,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,data.frame,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,data.frame,character
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,list,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,list,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,list,character
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,missing,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,missing,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,missing,character
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,character,missing
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,character,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for Collection,character,character
MbioDataset(collections, metadata, ontology)

# S4 method for list,SampleMetadata,missing
MbioDataset(collections, metadata, ontology)

# S4 method for list,SampleMetadata,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for list,SampleMetadata,character
MbioDataset(collections, metadata, ontology)

# S4 method for list,data.frame,missing
MbioDataset(collections, metadata, ontology)

# S4 method for list,data.frame,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for list,data.frame,character
MbioDataset(collections, metadata, ontology)

# S4 method for list,missing,missing
MbioDataset(collections, metadata, ontology)

# S4 method for list,missing,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for list,missing,character
MbioDataset(collections, metadata, ontology)

# S4 method for list,list,missing
MbioDataset(collections, metadata, ontology)

# S4 method for list,list,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for list,list,character
MbioDataset(collections, metadata, ontology)

# S4 method for list,character,missing
MbioDataset(collections, metadata, ontology)

# S4 method for list,character,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for list,character,character
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,SampleMetadata,missing
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,SampleMetadata,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,SampleMetadata,character
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,list,missing
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,list,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,list,character
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,missing,missing
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,missing,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,missing,character
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,data.frame,missing
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,data.frame,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,data.frame,character
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,character,missing
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,character,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for data.frame,character,character
MbioDataset(collections, metadata, ontology)

# S4 method for character,character,missing
MbioDataset(collections, metadata, ontology)

# S4 method for character,character,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for character,character,character
MbioDataset(collections, metadata, ontology)

# S4 method for character,list,missing
MbioDataset(collections, metadata, ontology)

# S4 method for character,list,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for character,list,character
MbioDataset(collections, metadata, ontology)

# S4 method for character,missing,missing
MbioDataset(collections, metadata, ontology)

# S4 method for character,missing,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for character,missing,character
MbioDataset(collections, metadata, ontology)

# S4 method for character,data.frame,missing
MbioDataset(collections, metadata, ontology)

# S4 method for character,data.frame,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for character,data.frame,character
MbioDataset(collections, metadata, ontology)

# S4 method for character,SampleMetadata,missing
MbioDataset(collections, metadata, ontology)

# S4 method for character,SampleMetadata,data.frame
MbioDataset(collections, metadata, ontology)

# S4 method for character,SampleMetadata,character
MbioDataset(collections, metadata, ontology)

Arguments

collections

A list of Collection objects, a data.frame containing multiple collections, or a character vector containing one or more file path(s)

metadata

A SampleMetadata object, a data.frame containing sample metadata, or a list of file path(s)

ontology

An data.frame containing the ontology of the dataset, or a character vector containing a file path to a data.frame