Maaslin2 from a Collection

MaAsLin2 finds associations between microbiome meta-omics features and complex metadata in population-scale epidemiological studies. The software includes multiple analysis methods (including support for multiple covariates and repeated measures) filtering, normalization, and transform options to customize analysis for your specific study.

Usage

Maaslin2(data, verbose = c(TRUE, FALSE), ...)

# S4 method for CollectionWithMetadata
Maaslin2(data, verbose = c(TRUE, FALSE), ...)

Arguments

data

a CollectionWithMetadata

verbose

boolean indicating if timed logging is desired

...

additional arguments to pass to Maaslin2::Maaslin2

Examples

maaslinOutput <- MicrobiomeDB::Maaslin2(
       data = getCollection(microbiomeData::DiabImmune, '16S (V4) Genus'), 
       output = tempfile("maaslin"),
       #min_prevalence = 0,
       fixed_effects = 'delivery_mode',
       analysis_method = "LM", # default LM
       normalization = "TSS", # default TSS
       transform = "LOG", # default LOG
       plot_heatmap = FALSE,
       plot_scatter = FALSE)
#> [1] "data.table" "data.frame"
#> [1] "Creating output folder"
#> [1] "Creating output feature tables folder"
#> [1] "Creating output fits folder"
#> 2024-04-24 17:16:17.137393 INFO::Writing function arguments to log file
#> 2024-04-24 17:16:17.152308 INFO::Verifying options selected are valid
#> 2024-04-24 17:16:17.181968 INFO::Determining format of input files
#> 2024-04-24 17:16:17.183541 INFO::Input format is data samples as rows and metadata samples as rows
#> 2024-04-24 17:16:17.197836 INFO::Formula for fixed effects: expr ~  delivery_mode
#> 2024-04-24 17:16:17.198775 INFO::Filter data based on min abundance and min prevalence
#> 2024-04-24 17:16:17.199319 INFO::Total samples in data: 3184
#> 2024-04-24 17:16:17.199795 INFO::Min samples required with min abundance for a feature not to be filtered: 318.400000
#> 2024-04-24 17:16:17.219658 INFO::Total filtered features: 563
#> 2024-04-24 17:16:17.220761 INFO::Filtered feature names from abundance and prevalence filtering: X28.4, Abiotrophia, Acetanaerobacterium, Acetobacter, Achromobacter, Acidaminococcus, Acidibacter, Acidiphilium, Acidipropionibacterium, Acidocella, Acidovorax, Acinetobacter, Actinobacillus, Actinobaculum, Actinotignum, Adlercreutzia, Aerococcus, Aeromonas, Afipia, Aggregatibacter, Alcaligenes, Alicycliphilus, Allisonella, Allobaculum, Alloprevotella, Allorhizobium.Neorhizobium.Pararhizobium.Rhizobium, Alloscardovia, Amaricoccus, Amnibacterium, Anaerobium, Anaerocolumna, Anaerofilum, Anaerofustis, Anaeroglobus, Anaeromyxobacter, Anaeroplasma, Anaerosporobacter, Anaerosporomusa, Anaerostignum, Anaerotruncus, Angelakisella, Angustibacter, Anoxybacillus, Aquabacterium, Aquamonas, Arcanobacterium, Arcticibacter, Asaccharospora, Asaia, Asinibacterium, Asteroleplasma, Asticcacaulis, Aureimonas, Azospira, Bacillus, Bacteriovorax, Bdellovibrio, Belnapia, Bergeyella, Blastomonas, Bordetella, Bosea, Brachybacterium, Bradyrhizobium, Brevibacterium, Brevundimonas, Brucella, Bulleidia, Burkholderia.Caballeronia.Paraburkholderia, Buttiauxella, Butyricimonas, CAG.56, CAG.873, CHKCI001, CHKCI002, CL500.29.marine.group, Caloribacterium, Candidatus.Arthromitus, Candidatus.Metachlamydia, Candidatus.Obscuribacter, Candidatus.Phytoplasma, Candidatus.Soleaferrea, Candidatus.Stoquefichus, Capnocytophaga, Caproiciproducens, Cardiobacterium, Carnobacterium, Caryophanon, Castellaniella, Catabacter, Catenibacillus, Catenibacterium, Catenisphaera, Catonella, Caulobacter, Cellulomonas, Cellulosilyticum, Cellvibrio, Centipeda, Cetobacterium, Christensenella, Chryseobacterium, Chryseomicrobium, Chungangia, Citrobacter, Clavibacter, Cloacibacillus, Cloacibacterium, Clostridium.sensu.stricto.12, Clostridium.sensu.stricto.13, Clostridium.sensu.stricto.15, Clostridium.sensu.stricto.18, Clostridium.sensu.stricto.2, Clostridium.sensu.stricto.3, Clostridium.sensu.stricto.4, Clostridium.sensu.stricto.6, Comamonas, Conexibacter, Conservatibacter, Coprobacter, Coriobacteriaceae.UCG.002, Corynebacterium, Cosenzaea, Criibacterium, Cryptobacterium, Curvibacter, Cutibacterium, Cytophaga, DNF00809, DSSD61, Dechloromonas, Defluviicoccus, Defluviitaleaceae.UCG.011, Delftia, Denitrobacterium, Dermabacter, Desemzia, Desulfovibrio, Devosia, Diaphorobacter, Dickeya, Dielma, Dietzia, Dinghuibacter, Diplosphaera, Dolosigranulum, Duganella, Dysgonomonas, Edaphobaculum, Eikenella, Empedobacter, Endobacter, Enhydrobacter, Enterobacter, Enterorhabdus, Enteroscipio, Epulopiscium, Erwinia, Erysipelotrichaceae.UCG.006, Eubacterium, Exiguobacterium, Extensimonas, F0058, F0332, FD2005, Facklamia, Faecalicoccus, Faecalitalea, Falsochrobactrum, Family.XIII.UCG.001, Fictibacillus, Flaviflexus, Flavihumibacter, Flavobacterium, Floricoccus, Fluviicola, Fournierella, Fretibacterium, Frisingicoccus, Fructobacillus, GCA.900066575, GCA.900066755, Galbibacter, Garciella, Gemmata, Geobacillus, Gluconobacter, Glutamicibacter, Gordonia, Gordonibacter, Gulbenkiania, Hafnia.Obesumbacterium, Halomonas, Haoranjiania, Harryflintia, Helicobacter, Hephaestia, Herbaspirillum, Herbinix, Holdemanella, Howardella, Hydrocarboniphaga, Hydrogenophilus, Hymenobacter, Iamia, Ignavigranum, Incertae.Sedis..EUPATH_0009256_Bacteria_Firmicutes_Clostridia_Lachnospirales_Lachnospiraceae_Incertae_Sedis., Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Erwiniaceae_Incertae_Sedis., Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Morganellaceae_Incertae_Sedis., Janthinobacterium, Jeotgalicoccus, Johnsonella, Kingella, Knoellia, Kocuria, Kosakonia, Lachnoanaerobaculum, Lachnobacterium, Lachnospiraceae.AC2044.group, Lachnospiraceae.FE2018.group, Lachnospiraceae.NC2004.group, Lachnospiraceae.UCG.002, Lachnospiraceae.UCG.003, Lachnospiraceae.UCG.008, Lachnospiraceae.XPB1014.group, Lachnotalea, Lactonifactor, Lautropia, Lawsonella, Legionella, Leptotrichia, Leuconostoc, Libanicoccus, Luteimonas, Luteolibacter, Macellibacteroides, Macrococcus, Mailhella, Marinilactibacillus, Marmoricola, Martelella, Marvinbryantia, Massilia, Megamonas, Merdibacter, Mesorhizobium, Methanobrevibacter, Methylibium, Methylobacillus, Methylobacterium.Methylorubrum, Methylorosula, Methylotenera, Micrococcus, Mitsuokella, Mobiluncus, Mogibacterium, Moheibacter, Mongoliitalea, Moraxella, Morganella, Moryella, Mucilaginibacter, Muribaculum, Murimonas, Mycobacterium, Mycoplasma, Myxococcus, NK4A214.group, Nakamurella, Nannocystis, Necropsobacter, Negativibacillus, Negativicoccus, Neochlamydia, Neomegalonema, Neorhizobium, Neoscardovia, Nesterenkonia, Nocardioides, Nocardiopsis, Novosphingobium, Oceanivirga, Oceanobacillus, Ochrobactrum, Olsenella, Oribacterium, Ottowia, Oxalobacter, Oxobacter, P3OB.42, Paenibacillus, Paeniclostridium, Paenisporosarcina, Paludicola, Papillibacter, Paraclostridium, Paracoccus, Paramesorhizobium, Paraprevotella, Parapusillimonas, Parascardovia, Parvibacter, Pasteurella, Paucibacter, Pediococcus, Pedobacter, Peptoclostridium, Peptococcus, Peptostreptococcus, Perlucidibaca, Phascolarctobacterium, Phaseolibacter, Phocea, Phoenicibacter, Phyllobacterium, Piscinibacter, Planifilum, Plesiomonas, Plot4.2H12, Pluralibacter, Polaromonas, Porphyromonas, Prevotellaceae.NK3B31.group, Prevotellaceae.UCG.001, Propionibacterium, Prosthecobacter, Proteus, Providencia, Pseudacidovorax, Pseudaminobacter, Pseudocitrobacter, Pseudoflavonifractor, Pseudoglutamicibacter, Pseudoramibacter, Pseudoxanthomonas, Psychrobacter, Pygmaiobacter, Pyramidobacter, Qipengyuania, Ralstonia, Raoultella, Raoultibacter, Reyranella, Rhizorhapis, Rhodanobacter, Rhodobacter, Rhodococcus, Rikenella, Rikenellaceae.RC9.gut.group, Rivicola, Robinsoniella, Rodentibacter, Roseisolibacter, Rubellimicrobium, Rummeliibacillus, S5.A14a, SH.PL14, Saccharothrix, Salinimicrobium, Salinisphaera, Salipaludibacillus, Salmonella, Sanguibacteroides, Sarcina, Scardovia, Sedimentibacter, Segetibacter, Selenomonas, Senegalimassilia, Serratia, Shimwellia, Shuttleworthia, Slackia, Sneathia, Solibacillus, Sphingoaurantiacus, Sphingobacterium, Sphingobium, Sphingomonas, Sphingopyxis, Sporolactobacillus, Stenotrophomonas, Stomatobaculum, Streptobacillus, Succinatimonas, Succiniclasticum, Succinivibrio, Sulfurospirillum, Synergistes, TM7x, Tannerella, Tatumella, Tepidimonas, Tepidiphilus, Tetragenococcus, Thermus, Timonella, Tolumonas, Trabulsiella, Tropheryma, Trueperella, UC5.1.2E3, UCG.004, UCG.007, UCG.008, UCG.009, Ureibacillus, Varibaculum, Variovorax, Verticiella, Victivallis, Vulcaniibacterium, Weissella, Xanthomonas, Yersinia, Yokenella, Zoogloea, X.Bacteroides..pectinophilus.group, X.Eubacterium..brachy.group, X.Eubacterium..nodatum.group, X.Eubacterium..oxidoreducens.group, X.Eubacterium..ruminantium.group, unclassified.0319.6G20, unclassified.67.14, unclassified.A0839, unclassified.Acetobacteraceae, unclassified.Acidaminococcaceae, unclassified.Actinobacteria, unclassified.Actinomycetaceae, unclassified.Aeromonadales, unclassified.Anaerococcus, unclassified.Anaerosalibacter, unclassified.Anaerovoracaceae, unclassified.Armatimonadales, unclassified.Atopobiaceae, unclassified.Bacillaceae, unclassified.Bacillales, unclassified.Bacilli, unclassified.Bacteroidales, unclassified.Bacteroidia, unclassified.Barnesiellaceae, unclassified.Bifidobacteriaceae, unclassified.Burkholderiales, unclassified.Butyricicoccaceae, unclassified.Caloramatoraceae, unclassified.Carnobacteriaceae, unclassified.Caulobacteraceae, unclassified.Christensenellaceae, unclassified.Chroococcidiopsaceae, unclassified.Clostridiaceae, unclassified.Clostridiales, unclassified.Comamonadaceae, unclassified.Coriobacteriales.Incertae.Sedis, unclassified.Coriobacteriales, unclassified.Corynebacteriaceae, unclassified.Desulfovibrionaceae, unclassified.Devosiaceae, unclassified.Diplorickettsiaceae, unclassified.Eggerthellaceae, unclassified.Enterobacterales, unclassified.Enterobacteriaceae, unclassified.Enterococcaceae, unclassified.Erwiniaceae, unclassified.Erysipelatoclostridiaceae, unclassified.Erysipelotrichaceae, unclassified.Erysipelotrichales, unclassified.Ethanoligenenaceae, unclassified.Eubacteriaceae, unclassified.Ezakiella, unclassified.Fastidiosipila, unclassified.Fenollaria, unclassified.Fimbriimonadaceae, unclassified.Finegoldia, unclassified.Firmicutes, unclassified.Flavobacteriaceae, unclassified.Flavobacteriales, unclassified.Fusobacteriaceae, unclassified.Gaiellales, unclassified.Gallicola, unclassified.Gammaproteobacteria, unclassified.Gemmataceae, unclassified.Hungateiclostridiaceae, unclassified.Hydrogenoanaerobacterium, unclassified.Izemoplasmatales, unclassified.JI49D030, unclassified.KD3.93, unclassified.Kapabacteriales, unclassified.Lachnospirales, unclassified.Lactobacillales, unclassified.Leuconostocaceae, unclassified.Lineage.IV, unclassified.MAT.CR.H4.C10, unclassified.Methylococcaceae, unclassified.Micavibrionales, unclassified.Microbacteriaceae, unclassified.Micrococcaceae, unclassified.Micrococcales, unclassified.Moraxellaceae, unclassified.Murdochiella, unclassified.Muribaculaceae, unclassified.NS9.marine.group, unclassified.Neisseriaceae, unclassified.OLB14, unclassified.Opitutales, unclassified.Oscillospirales, unclassified.Oxalobacteraceae, unclassified.P5D1.392, unclassified.PLTA13, unclassified.Paenibacillaceae, unclassified.Paracaedibacteraceae, unclassified.Parachlamydiaceae, unclassified.Parvimonas, unclassified.Pasteurellaceae, unclassified.Peptococcaceae, unclassified.Peptostreptococcaceae, unclassified.Peptostreptococcales.Tissierellales, unclassified.Planctomycetales, unclassified.Planococcaceae, unclassified.Prevotellaceae, unclassified.Promicromonosporaceae, unclassified.Propionibacteriaceae, unclassified.Proteobacteria, unclassified.Pseudomonadaceae, unclassified.Puniceicoccaceae, unclassified.RF39, unclassified.Rhizobiaceae, unclassified.Rhizobiales.Incertae.Sedis, unclassified.Rhizobiales, unclassified.Rhodobacteraceae, unclassified.Rhodocyclaceae, unclassified.Rickettsiales, unclassified.Rikenellaceae, unclassified.Ruminiclostridium, unclassified.ST.12K33, unclassified.Saccharimonadales, unclassified.Sandaracinaceae, unclassified.Selenomonadaceae, unclassified.Simkaniaceae, unclassified.Solimonadaceae, unclassified.Sphingobacteriales, unclassified.Sphingomonadaceae, unclassified.Staphylococcaceae, unclassified.Streptococcaceae, unclassified.Succinivibrionaceae, unclassified.Sutterellaceae, unclassified.TRA3.20, unclassified.Tannerellaceae, unclassified.Tepidimicrobium, unclassified.UCG.010, unclassified.UCG.011, unclassified.Veillonellaceae, unclassified.Veillonellales.Selenomonadales, unclassified.Vermiphilaceae, unclassified.Verrucomicrobiales, unclassified.Verrucomicrobiota, unclassified.Victivallaceae, unclassified.W5053, unclassified.WPS.2, unclassified.Weeksellaceae, unclassified.Xanthobacteraceae, unclassified.Yersiniaceae, unclassified..Clostridium..methylpentosum.group, unclassified.vadinBE97
#> 2024-04-24 17:16:17.228667 INFO::Total filtered features with variance filtering: 0
#> 2024-04-24 17:16:17.229223 INFO::Filtered feature names from variance filtering:
#> 2024-04-24 17:16:17.229697 INFO::Running selected normalization method: TSS
#> 2024-04-24 17:16:17.288467 INFO::Applying z-score to standardize continuous metadata
#> 2024-04-24 17:16:17.296834 INFO::Running selected transform method: LOG
#> 2024-04-24 17:16:17.35947 INFO::Running selected analysis method: LM
#> 2024-04-24 17:16:17.365472 INFO::Fitting model to feature number 1, Actinomyces
#> 2024-04-24 17:16:17.374012 INFO::Fitting model to feature number 2, Agathobacter
#> 2024-04-24 17:16:17.378144 INFO::Fitting model to feature number 3, Akkermansia
#> 2024-04-24 17:16:17.38224 INFO::Fitting model to feature number 4, Alistipes
#> 2024-04-24 17:16:17.386331 INFO::Fitting model to feature number 5, Anaerostipes
#> 2024-04-24 17:16:17.390452 INFO::Fitting model to feature number 6, Atopobium
#> 2024-04-24 17:16:17.3945 INFO::Fitting model to feature number 7, Bacteroides
#> 2024-04-24 17:16:17.398598 INFO::Fitting model to feature number 8, Barnesiella
#> 2024-04-24 17:16:17.402634 INFO::Fitting model to feature number 9, Bifidobacterium
#> 2024-04-24 17:16:17.406713 INFO::Fitting model to feature number 10, Bilophila
#> 2024-04-24 17:16:17.410826 INFO::Fitting model to feature number 11, Blautia
#> 2024-04-24 17:16:17.432689 INFO::Fitting model to feature number 12, Butyricicoccus
#> 2024-04-24 17:16:17.437162 INFO::Fitting model to feature number 13, CAG.352
#> 2024-04-24 17:16:17.441193 INFO::Fitting model to feature number 14, Campylobacter
#> 2024-04-24 17:16:17.445297 INFO::Fitting model to feature number 15, Christensenellaceae.R.7.group
#> 2024-04-24 17:16:17.449362 INFO::Fitting model to feature number 16, Clostridioides
#> 2024-04-24 17:16:17.453418 INFO::Fitting model to feature number 17, Clostridium.sensu.stricto.1
#> 2024-04-24 17:16:17.457496 INFO::Fitting model to feature number 18, Colidextribacter
#> 2024-04-24 17:16:17.46159 INFO::Fitting model to feature number 19, Collinsella
#> 2024-04-24 17:16:17.465642 INFO::Fitting model to feature number 20, Coprobacillus
#> 2024-04-24 17:16:17.46976 INFO::Fitting model to feature number 21, Coprococcus
#> 2024-04-24 17:16:17.473836 INFO::Fitting model to feature number 22, DTU089
#> 2024-04-24 17:16:17.477946 INFO::Fitting model to feature number 23, Dialister
#> 2024-04-24 17:16:17.482205 INFO::Fitting model to feature number 24, Dorea
#> 2024-04-24 17:16:17.486331 INFO::Fitting model to feature number 25, Eggerthella
#> 2024-04-24 17:16:17.490447 INFO::Fitting model to feature number 26, Eisenbergiella
#> 2024-04-24 17:16:17.494556 INFO::Fitting model to feature number 27, Enterococcus
#> 2024-04-24 17:16:17.498703 INFO::Fitting model to feature number 28, Erysipelatoclostridium
#> 2024-04-24 17:16:17.502821 INFO::Fitting model to feature number 29, Erysipelotrichaceae.UCG.003
#> 2024-04-24 17:16:17.506915 INFO::Fitting model to feature number 30, Escherichia.Shigella
#> 2024-04-24 17:16:17.510962 INFO::Fitting model to feature number 31, Faecalibacterium
#> 2024-04-24 17:16:17.515036 INFO::Fitting model to feature number 32, Family.XIII.AD3011.group
#> 2024-04-24 17:16:17.519112 INFO::Fitting model to feature number 33, Flavonifractor
#> 2024-04-24 17:16:17.523153 INFO::Fitting model to feature number 34, Fusicatenibacter
#> 2024-04-24 17:16:17.527172 INFO::Fitting model to feature number 35, Fusobacterium
#> 2024-04-24 17:16:17.531253 INFO::Fitting model to feature number 36, Gemella
#> 2024-04-24 17:16:17.535288 INFO::Fitting model to feature number 37, Granulicatella
#> 2024-04-24 17:16:17.539339 INFO::Fitting model to feature number 38, Haemophilus
#> 2024-04-24 17:16:17.543461 INFO::Fitting model to feature number 39, Holdemania
#> 2024-04-24 17:16:17.547532 INFO::Fitting model to feature number 40, Hungatella
#> 2024-04-24 17:16:17.551624 INFO::Fitting model to feature number 41, Incertae.Sedis..EUPATH_0009256_Bacteria_Firmicutes_Clostridia_Oscillospirales_Ruminococcaceae_Incertae_Sedis.
#> 2024-04-24 17:16:17.555752 INFO::Fitting model to feature number 42, Intestinibacter
#> 2024-04-24 17:16:17.567054 INFO::Fitting model to feature number 43, Intestinimonas
#> 2024-04-24 17:16:17.571323 INFO::Fitting model to feature number 44, Klebsiella
#> 2024-04-24 17:16:17.575483 INFO::Fitting model to feature number 45, Lachnoclostridium
#> 2024-04-24 17:16:17.579586 INFO::Fitting model to feature number 46, Lachnospira
#> 2024-04-24 17:16:17.58366 INFO::Fitting model to feature number 47, Lachnospiraceae.FCS020.group
#> 2024-04-24 17:16:17.587696 INFO::Fitting model to feature number 48, Lachnospiraceae.ND3007.group
#> 2024-04-24 17:16:17.591765 INFO::Fitting model to feature number 49, Lachnospiraceae.NK4A136.group
#> 2024-04-24 17:16:17.595816 INFO::Fitting model to feature number 50, Lachnospiraceae.UCG.001
#> 2024-04-24 17:16:17.599863 INFO::Fitting model to feature number 51, Lachnospiraceae.UCG.004
#> 2024-04-24 17:16:17.603962 INFO::Fitting model to feature number 52, Lachnospiraceae.UCG.009
#> 2024-04-24 17:16:17.608043 INFO::Fitting model to feature number 53, Lachnospiraceae.UCG.010
#> 2024-04-24 17:16:17.612105 INFO::Fitting model to feature number 54, Lactobacillus
#> 2024-04-24 17:16:17.616189 INFO::Fitting model to feature number 55, Lactococcus
#> 2024-04-24 17:16:17.620257 INFO::Fitting model to feature number 56, Mannheimia
#> 2024-04-24 17:16:17.624327 INFO::Fitting model to feature number 57, Megasphaera
#> 2024-04-24 17:16:17.628387 INFO::Fitting model to feature number 58, Monoglobus
#> 2024-04-24 17:16:17.632524 INFO::Fitting model to feature number 59, Neisseria
#> 2024-04-24 17:16:17.636647 INFO::Fitting model to feature number 60, Odoribacter
#> 2024-04-24 17:16:17.640717 INFO::Fitting model to feature number 61, Oscillibacter
#> 2024-04-24 17:16:17.644774 INFO::Fitting model to feature number 62, Oscillospira
#> 2024-04-24 17:16:17.64883 INFO::Fitting model to feature number 63, Parabacteroides
#> 2024-04-24 17:16:17.652919 INFO::Fitting model to feature number 64, Parasutterella
#> 2024-04-24 17:16:17.657008 INFO::Fitting model to feature number 65, Prevotella
#> 2024-04-24 17:16:17.661202 INFO::Fitting model to feature number 66, Pseudomonas
#> 2024-04-24 17:16:17.665254 INFO::Fitting model to feature number 67, Romboutsia
#> 2024-04-24 17:16:17.669317 INFO::Fitting model to feature number 68, Roseburia
#> 2024-04-24 17:16:17.673414 INFO::Fitting model to feature number 69, Rothia
#> 2024-04-24 17:16:17.677464 INFO::Fitting model to feature number 70, Ruminococcus
#> 2024-04-24 17:16:17.681498 INFO::Fitting model to feature number 71, Sellimonas
#> 2024-04-24 17:16:17.685598 INFO::Fitting model to feature number 72, Solobacterium
#> 2024-04-24 17:16:17.689627 INFO::Fitting model to feature number 73, Staphylococcus
#> 2024-04-24 17:16:17.701293 INFO::Fitting model to feature number 74, Streptococcus
#> 2024-04-24 17:16:17.705551 INFO::Fitting model to feature number 75, Subdoligranulum
#> 2024-04-24 17:16:17.709659 INFO::Fitting model to feature number 76, Sutterella
#> 2024-04-24 17:16:17.713776 INFO::Fitting model to feature number 77, Terrisporobacter
#> 2024-04-24 17:16:17.717878 INFO::Fitting model to feature number 78, Turicibacter
#> 2024-04-24 17:16:17.721924 INFO::Fitting model to feature number 79, Tuzzerella
#> 2024-04-24 17:16:17.72599 INFO::Fitting model to feature number 80, Tyzzerella
#> 2024-04-24 17:16:17.730034 INFO::Fitting model to feature number 81, UBA1819
#> 2024-04-24 17:16:17.734133 INFO::Fitting model to feature number 82, UCG.002
#> 2024-04-24 17:16:17.738212 INFO::Fitting model to feature number 83, UCG.003
#> 2024-04-24 17:16:17.742338 INFO::Fitting model to feature number 84, UCG.005
#> 2024-04-24 17:16:17.746497 INFO::Fitting model to feature number 85, Veillonella
#> 2024-04-24 17:16:17.750615 INFO::Fitting model to feature number 86, X.Clostridium..innocuum.group
#> 2024-04-24 17:16:17.754696 INFO::Fitting model to feature number 87, X.Eubacterium..eligens.group
#> 2024-04-24 17:16:17.758765 INFO::Fitting model to feature number 88, X.Eubacterium..fissicatena.group
#> 2024-04-24 17:16:17.762878 INFO::Fitting model to feature number 89, X.Eubacterium..hallii.group
#> 2024-04-24 17:16:17.767056 INFO::Fitting model to feature number 90, X.Eubacterium..siraeum.group
#> 2024-04-24 17:16:17.771185 INFO::Fitting model to feature number 91, X.Eubacterium..ventriosum.group
#> 2024-04-24 17:16:17.775296 INFO::Fitting model to feature number 92, X.Eubacterium..xylanophilum.group
#> 2024-04-24 17:16:17.779359 INFO::Fitting model to feature number 93, X.Ruminococcus..gauvreauii.group
#> 2024-04-24 17:16:17.783436 INFO::Fitting model to feature number 94, X.Ruminococcus..gnavus.group
#> 2024-04-24 17:16:17.787523 INFO::Fitting model to feature number 95, X.Ruminococcus..torques.group
#> 2024-04-24 17:16:17.79159 INFO::Fitting model to feature number 96, unclassified.Chloroplast
#> 2024-04-24 17:16:17.795679 INFO::Fitting model to feature number 97, unclassified.Clostridia.UCG.014
#> 2024-04-24 17:16:17.799751 INFO::Fitting model to feature number 98, unclassified.Clostridia
#> 2024-04-24 17:16:17.803801 INFO::Fitting model to feature number 99, unclassified.Clostridia.vadinBB60.group
#> 2024-04-24 17:16:17.807838 INFO::Fitting model to feature number 100, unclassified.Gastranaerophilales
#> 2024-04-24 17:16:17.811925 INFO::Fitting model to feature number 101, unclassified.Lachnospiraceae
#> 2024-04-24 17:16:17.815993 INFO::Fitting model to feature number 102, unclassified.Mitochondria
#> 2024-04-24 17:16:17.827484 INFO::Fitting model to feature number 103, unclassified.Oscillospiraceae
#> 2024-04-24 17:16:17.831639 INFO::Fitting model to feature number 104, unclassified.Peptoniphilus
#> 2024-04-24 17:16:17.835698 INFO::Fitting model to feature number 105, unclassified.Rhodospirillales
#> 2024-04-24 17:16:17.839779 INFO::Fitting model to feature number 106, unclassified.Ruminococcaceae
#> 2024-04-24 17:16:17.843884 INFO::Fitting model to feature number 107, unclassified..Eubacterium..coprostanoligenes.group
#> 2024-04-24 17:16:17.863447 INFO::Counting total values for each feature
#> 2024-04-24 17:16:17.869231 INFO::Writing filtered data to file /tmp/RtmpfBGTvS/maaslin195842d2cad5/features/filtered_data.tsv
#> 2024-04-24 17:16:18.014656 INFO::Writing filtered, normalized data to file /tmp/RtmpfBGTvS/maaslin195842d2cad5/features/filtered_data_norm.tsv
#> 2024-04-24 17:16:18.16445 INFO::Writing filtered, normalized, transformed data to file /tmp/RtmpfBGTvS/maaslin195842d2cad5/features/filtered_data_norm_transformed.tsv
#> 2024-04-24 17:16:18.437835 INFO::Writing residuals to file /tmp/RtmpfBGTvS/maaslin195842d2cad5/fits/residuals.rds
#> 2024-04-24 17:16:18.512103 INFO::Writing fitted values to file /tmp/RtmpfBGTvS/maaslin195842d2cad5/fits/fitted.rds
#> 2024-04-24 17:16:18.536584 INFO::Writing all results to file (ordered by increasing q-values): /tmp/RtmpfBGTvS/maaslin195842d2cad5/all_results.tsv
#> 2024-04-24 17:16:18.538183 INFO::Writing the significant results (those which are less than or equal to the threshold of 0.250000 ) to file (ordered by increasing q-values): /tmp/RtmpfBGTvS/maaslin195842d2cad5/significant_results.tsv