Maaslin2 from a Collection

MaAsLin2 finds associations between microbiome meta-omics features and complex metadata in population-scale epidemiological studies. The software includes multiple analysis methods (including support for multiple covariates and repeated measures) filtering, normalization, and transform options to customize analysis for your specific study.

Usage

Maaslin2(data, verbose = c(TRUE, FALSE), ...)

# S4 method for CollectionWithMetadata
Maaslin2(data, verbose = c(TRUE, FALSE), ...)

Arguments

data

a CollectionWithMetadata

verbose

boolean indicating if timed logging is desired

...

additional arguments to pass to Maaslin2::Maaslin2

Details

The MicrobiomeDB wrapper for Maaslin2 will ignore samples where the fixed_effects variable has a value of NA or the empty string ("").

Examples

maaslinOutput <- MicrobiomeDB::Maaslin2(
       data = getCollection(
  microbiomeData::DiabImmune, 
  '16S (V4) Genus (Relative taxonomic abundance analysis)'), 
       output = tempfile("maaslin"),
       #min_prevalence = 0,
       fixed_effects = 'delivery_mode',
       analysis_method = "LM", # default LM
       normalization = "TSS", # default TSS
       transform = "LOG", # default LOG
       plot_heatmap = FALSE,
       plot_scatter = FALSE)
#> [1] "Creating output folder"
#> [1] "Creating output feature tables folder"
#> [1] "Creating output fits folder"
#> 2024-05-29 18:02:11.459499 INFO::Writing function arguments to log file
#> 2024-05-29 18:02:11.474672 INFO::Verifying options selected are valid
#> 2024-05-29 18:02:11.503937 INFO::Determining format of input files
#> 2024-05-29 18:02:11.505837 INFO::Input format is data samples as rows and metadata samples as rows
#> 2024-05-29 18:02:11.521223 INFO::Formula for fixed effects: expr ~  delivery_mode
#> 2024-05-29 18:02:11.522336 INFO::Filter data based on min abundance and min prevalence
#> 2024-05-29 18:02:11.522908 INFO::Total samples in data: 3184
#> 2024-05-29 18:02:11.52339 INFO::Min samples required with min abundance for a feature not to be filtered: 318.400000
#> 2024-05-29 18:02:11.555755 INFO::Total filtered features: 564
#> 2024-05-29 18:02:11.556865 INFO::Filtered feature names from abundance and prevalence filtering: X28.4, Abiotrophia, Acetanaerobacterium, Acetobacter, Achromobacter, Acidaminococcus, Acidibacter, Acidiphilium, Acidipropionibacterium, Acidocella, Acidovorax, Acinetobacter, Actinobacillus, Actinobaculum, Actinotignum, Adlercreutzia, Aerococcus, Aeromonas, Afipia, Aggregatibacter, Alcaligenes, Alicycliphilus, Allisonella, Allobaculum, Alloprevotella, Allorhizobium.Neorhizobium.Pararhizobium.Rhizobium, Alloscardovia, Amaricoccus, Amnibacterium, Anaerobium, Anaerocolumna, Anaerofilum, Anaerofustis, Anaeroglobus, Anaeromyxobacter, Anaeroplasma, Anaerosporobacter, Anaerosporomusa, Anaerostignum, Anaerotruncus, Angelakisella, Angustibacter, Anoxybacillus, Aquabacterium, Aquamonas, Arcanobacterium, Arcticibacter, Asaccharospora, Asaia, Asinibacterium, Asteroleplasma, Asticcacaulis, Aureimonas, Azospira, Bacillus, Bacteriovorax, Bdellovibrio, Belnapia, Bergeyella, Blastomonas, Bordetella, Bosea, Brachybacterium, Bradyrhizobium, Brevibacterium, Brevundimonas, Brucella, Bulleidia, Burkholderia.Caballeronia.Paraburkholderia, Buttiauxella, Butyricimonas, CAG.56, CAG.873, CHKCI001, CHKCI002, CL500.29.marine.group, Caloribacterium, Candidatus.Arthromitus, Candidatus.Metachlamydia, Candidatus.Obscuribacter, Candidatus.Phytoplasma, Candidatus.Soleaferrea, Candidatus.Stoquefichus, Capnocytophaga, Caproiciproducens, Cardiobacterium, Carnobacterium, Caryophanon, Castellaniella, Catabacter, Catenibacillus, Catenibacterium, Catenisphaera, Catonella, Caulobacter, Cellulomonas, Cellulosilyticum, Cellvibrio, Centipeda, Cetobacterium, Christensenella, Chryseobacterium, Chryseomicrobium, Chungangia, Citrobacter, Clavibacter, Cloacibacillus, Cloacibacterium, Clostridium.sensu.stricto.12, Clostridium.sensu.stricto.13, Clostridium.sensu.stricto.15, Clostridium.sensu.stricto.18, Clostridium.sensu.stricto.2, Clostridium.sensu.stricto.3, Clostridium.sensu.stricto.4, Clostridium.sensu.stricto.6, Comamonas, Conexibacter, Conservatibacter, Coprobacter, Coriobacteriaceae.UCG.002, Corynebacterium, Cosenzaea, Criibacterium, Cryptobacterium, Curvibacter, Cutibacterium, Cytophaga, DNF00809, DSSD61, Dechloromonas, Defluviicoccus, Defluviitaleaceae.UCG.011, Delftia, Denitrobacterium, Dermabacter, Desemzia, Desulfovibrio, Devosia, Diaphorobacter, Dickeya, Dielma, Dietzia, Dinghuibacter, Diplosphaera, Dolosigranulum, Duganella, Dysgonomonas, Edaphobaculum, Eikenella, Empedobacter, Endobacter, Enhydrobacter, Enterobacter, Enterorhabdus, Enteroscipio, Epulopiscium, Erwinia, Erysipelotrichaceae.UCG.006, Eubacterium, Exiguobacterium, Extensimonas, F0058, F0332, FD2005, Facklamia, Faecalicoccus, Faecalitalea, Falsochrobactrum, Family.XIII.UCG.001, Fictibacillus, Flaviflexus, Flavihumibacter, Flavobacterium, Floricoccus, Fluviicola, Fournierella, Fretibacterium, Frisingicoccus, Fructobacillus, GCA.900066575, GCA.900066755, Galbibacter, Garciella, Gemmata, Geobacillus, Gluconobacter, Glutamicibacter, Gordonia, Gordonibacter, Gulbenkiania, Hafnia.Obesumbacterium, Halomonas, Haoranjiania, Harryflintia, Helicobacter, Hephaestia, Herbaspirillum, Herbinix, Holdemanella, Howardella, Hydrocarboniphaga, Hydrogenophilus, Hymenobacter, Iamia, Ignavigranum, Incertae.Sedis..EUPATH_0009256_Bacteria_Firmicutes_Clostridia_Lachnospirales_Lachnospiraceae_Incertae_Sedis., Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Erwiniaceae_Incertae_Sedis., Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Morganellaceae_Incertae_Sedis., Janthinobacterium, Jeotgalicoccus, Johnsonella, Kingella, Knoellia, Kocuria, Kosakonia, Lachnoanaerobaculum, Lachnobacterium, Lachnospiraceae.AC2044.group, Lachnospiraceae.FE2018.group, Lachnospiraceae.NC2004.group, Lachnospiraceae.UCG.002, Lachnospiraceae.UCG.003, Lachnospiraceae.UCG.008, Lachnospiraceae.XPB1014.group, Lachnotalea, Lactonifactor, Lautropia, Lawsonella, Legionella, Leptotrichia, Leuconostoc, Libanicoccus, Luteimonas, Luteolibacter, Macellibacteroides, Macrococcus, Mailhella, Marinilactibacillus, Marmoricola, Martelella, Marvinbryantia, Massilia, Megamonas, Merdibacter, Mesorhizobium, Methanobrevibacter, Methylibium, Methylobacillus, Methylobacterium.Methylorubrum, Methylorosula, Methylotenera, Micrococcus, Mitsuokella, Mobiluncus, Mogibacterium, Moheibacter, Mongoliitalea, Moraxella, Morganella, Moryella, Mucilaginibacter, Muribaculum, Murimonas, Mycobacterium, Mycoplasma, Myxococcus, NK4A214.group, Nakamurella, Nannocystis, Necropsobacter, Negativibacillus, Negativicoccus, Neochlamydia, Neomegalonema, Neorhizobium, Neoscardovia, Nesterenkonia, Nocardioides, Nocardiopsis, Novosphingobium, Oceanivirga, Oceanobacillus, Ochrobactrum, Olsenella, Oribacterium, Ottowia, Oxalobacter, Oxobacter, P3OB.42, Paenibacillus, Paeniclostridium, Paenisporosarcina, Paludicola, Papillibacter, Paraclostridium, Paracoccus, Paramesorhizobium, Paraprevotella, Parapusillimonas, Parascardovia, Parvibacter, Pasteurella, Paucibacter, Pediococcus, Pedobacter, Peptoclostridium, Peptococcus, Peptostreptococcus, Perlucidibaca, Phascolarctobacterium, Phaseolibacter, Phocea, Phoenicibacter, Phyllobacterium, Piscinibacter, Planifilum, Plesiomonas, Plot4.2H12, Pluralibacter, Polaromonas, Porphyromonas, Prevotellaceae.NK3B31.group, Prevotellaceae.UCG.001, Propionibacterium, Propionimicrobium, Prosthecobacter, Proteus, Providencia, Pseudacidovorax, Pseudaminobacter, Pseudocitrobacter, Pseudoflavonifractor, Pseudoglutamicibacter, Pseudoramibacter, Pseudoxanthomonas, Psychrobacter, Pygmaiobacter, Pyramidobacter, Qipengyuania, Ralstonia, Raoultella, Raoultibacter, Reyranella, Rhizorhapis, Rhodanobacter, Rhodobacter, Rhodococcus, Rikenella, Rikenellaceae.RC9.gut.group, Rivicola, Robinsoniella, Rodentibacter, Roseisolibacter, Rubellimicrobium, Rummeliibacillus, S5.A14a, SH.PL14, Saccharothrix, Salinimicrobium, Salinisphaera, Salipaludibacillus, Salmonella, Sanguibacteroides, Sarcina, Scardovia, Sedimentibacter, Segetibacter, Selenomonas, Senegalimassilia, Serratia, Shimwellia, Shuttleworthia, Slackia, Sneathia, Solibacillus, Sphingoaurantiacus, Sphingobacterium, Sphingobium, Sphingomonas, Sphingopyxis, Sporolactobacillus, Stenotrophomonas, Stomatobaculum, Streptobacillus, Succinatimonas, Succiniclasticum, Succinivibrio, Sulfurospirillum, Synergistes, TM7x, Tannerella, Tatumella, Tepidimonas, Tepidiphilus, Tetragenococcus, Thermus, Timonella, Tolumonas, Trabulsiella, Tropheryma, Trueperella, UC5.1.2E3, UCG.004, UCG.007, UCG.008, UCG.009, Ureibacillus, Varibaculum, Variovorax, Verticiella, Victivallis, Vulcaniibacterium, Weissella, Xanthomonas, Yersinia, Yokenella, Zoogloea, X.Bacteroides..pectinophilus.group, X.Eubacterium..brachy.group, X.Eubacterium..nodatum.group, X.Eubacterium..oxidoreducens.group, X.Eubacterium..ruminantium.group, unclassified.0319.6G20, unclassified.67.14, unclassified.A0839, unclassified.Acetobacteraceae, unclassified.Acidaminococcaceae, unclassified.Actinobacteria, unclassified.Actinomycetaceae, unclassified.Aeromonadales, unclassified.Anaerococcus, unclassified.Anaerosalibacter, unclassified.Anaerovoracaceae, unclassified.Armatimonadales, unclassified.Atopobiaceae, unclassified.Bacillaceae, unclassified.Bacillales, unclassified.Bacilli, unclassified.Bacteroidales, unclassified.Bacteroidia, unclassified.Barnesiellaceae, unclassified.Bifidobacteriaceae, unclassified.Burkholderiales, unclassified.Butyricicoccaceae, unclassified.Caloramatoraceae, unclassified.Carnobacteriaceae, unclassified.Caulobacteraceae, unclassified.Christensenellaceae, unclassified.Chroococcidiopsaceae, unclassified.Clostridiaceae, unclassified.Clostridiales, unclassified.Comamonadaceae, unclassified.Coriobacteriales.Incertae.Sedis..EUPATH_0009256_Bacteria_Actinobacteriota_Coriobacteriia_Coriobacteriales_Coriobacteriales_Incertae_Sedis_., unclassified.Coriobacteriales, unclassified.Corynebacteriaceae, unclassified.Desulfovibrionaceae, unclassified.Devosiaceae, unclassified.Diplorickettsiaceae, unclassified.Eggerthellaceae, unclassified.Enterobacterales, unclassified.Enterobacteriaceae, unclassified.Enterococcaceae, unclassified.Erwiniaceae, unclassified.Erysipelatoclostridiaceae, unclassified.Erysipelotrichaceae, unclassified.Erysipelotrichales, unclassified.Ethanoligenenaceae, unclassified.Eubacteriaceae, unclassified.Ezakiella, unclassified.Fastidiosipila, unclassified.Fenollaria, unclassified.Fimbriimonadaceae, unclassified.Finegoldia, unclassified.Firmicutes, unclassified.Flavobacteriaceae, unclassified.Flavobacteriales, unclassified.Fusobacteriaceae, unclassified.Gaiellales, unclassified.Gallicola, unclassified.Gammaproteobacteria, unclassified.Gemmataceae, unclassified.Hungateiclostridiaceae, unclassified.Hydrogenoanaerobacterium, unclassified.Izemoplasmatales, unclassified.JI49D030, unclassified.KD3.93, unclassified.Kapabacteriales, unclassified.Lachnospirales, unclassified.Lactobacillales, unclassified.Leuconostocaceae, unclassified.Lineage.IV, unclassified.MAT.CR.H4.C10, unclassified.Methylococcaceae, unclassified.Micavibrionales, unclassified.Microbacteriaceae, unclassified.Micrococcaceae, unclassified.Micrococcales, unclassified.Moraxellaceae, unclassified.Murdochiella, unclassified.Muribaculaceae, unclassified.NS9.marine.group, unclassified.Neisseriaceae, unclassified.OLB14, unclassified.Opitutales, unclassified.Oscillospirales, unclassified.Oxalobacteraceae, unclassified.P5D1.392, unclassified.PLTA13, unclassified.Paenibacillaceae, unclassified.Paracaedibacteraceae, unclassified.Parachlamydiaceae, unclassified.Parvimonas, unclassified.Pasteurellaceae, unclassified.Peptococcaceae, unclassified.Peptostreptococcaceae, unclassified.Peptostreptococcales.Tissierellales, unclassified.Planctomycetales, unclassified.Planococcaceae, unclassified.Prevotellaceae, unclassified.Promicromonosporaceae, unclassified.Propionibacteriaceae, unclassified.Proteobacteria, unclassified.Pseudomonadaceae, unclassified.Puniceicoccaceae, unclassified.RF39, unclassified.Rhizobiaceae, unclassified.Rhizobiales.Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Alphaproteobacteria_Rhizobiales_Rhizobiales_Incertae_Sedis_., unclassified.Rhizobiales, unclassified.Rhodobacteraceae, unclassified.Rhodocyclaceae, unclassified.Rickettsiales, unclassified.Rikenellaceae, unclassified.Ruminiclostridium, unclassified.ST.12K33, unclassified.Saccharimonadales, unclassified.Sandaracinaceae, unclassified.Selenomonadaceae, unclassified.Simkaniaceae, unclassified.Solimonadaceae, unclassified.Sphingobacteriales, unclassified.Sphingomonadaceae, unclassified.Staphylococcaceae, unclassified.Streptococcaceae, unclassified.Succinivibrionaceae, unclassified.Sutterellaceae, unclassified.TRA3.20, unclassified.Tannerellaceae, unclassified.Tepidimicrobium, unclassified.UCG.010, unclassified.UCG.011, unclassified.Veillonellaceae, unclassified.Veillonellales.Selenomonadales, unclassified.Vermiphilaceae, unclassified.Verrucomicrobiales, unclassified.Verrucomicrobiota, unclassified.Victivallaceae, unclassified.W5053, unclassified.WPS.2, unclassified.Weeksellaceae, unclassified.Xanthobacteraceae, unclassified.Yersiniaceae, unclassified..Clostridium..methylpentosum.group, unclassified.vadinBE97
#> 2024-05-29 18:02:11.56493 INFO::Total filtered features with variance filtering: 0
#> 2024-05-29 18:02:11.565623 INFO::Filtered feature names from variance filtering:
#> 2024-05-29 18:02:11.5661 INFO::Running selected normalization method: TSS
#> 2024-05-29 18:02:11.629113 INFO::Applying z-score to standardize continuous metadata
#> 2024-05-29 18:02:11.637299 INFO::Running selected transform method: LOG
#> 2024-05-29 18:02:11.711167 INFO::Running selected analysis method: LM
#> 2024-05-29 18:02:11.717269 INFO::Fitting model to feature number 1, Actinomyces
#> 2024-05-29 18:02:11.74693 INFO::Fitting model to feature number 2, Agathobacter
#> 2024-05-29 18:02:11.751701 INFO::Fitting model to feature number 3, Akkermansia
#> 2024-05-29 18:02:11.756187 INFO::Fitting model to feature number 4, Alistipes
#> 2024-05-29 18:02:11.760638 INFO::Fitting model to feature number 5, Anaerostipes
#> 2024-05-29 18:02:11.765072 INFO::Fitting model to feature number 6, Atopobium
#> 2024-05-29 18:02:11.769583 INFO::Fitting model to feature number 7, Bacteroides
#> 2024-05-29 18:02:11.774034 INFO::Fitting model to feature number 8, Barnesiella
#> 2024-05-29 18:02:11.778467 INFO::Fitting model to feature number 9, Bifidobacterium
#> 2024-05-29 18:02:11.782941 INFO::Fitting model to feature number 10, Bilophila
#> 2024-05-29 18:02:11.78741 INFO::Fitting model to feature number 11, Blautia
#> 2024-05-29 18:02:11.791871 INFO::Fitting model to feature number 12, Butyricicoccus
#> 2024-05-29 18:02:11.796304 INFO::Fitting model to feature number 13, CAG.352
#> 2024-05-29 18:02:11.800789 INFO::Fitting model to feature number 14, Campylobacter
#> 2024-05-29 18:02:11.805244 INFO::Fitting model to feature number 15, Christensenellaceae.R.7.group
#> 2024-05-29 18:02:11.809726 INFO::Fitting model to feature number 16, Clostridioides
#> 2024-05-29 18:02:11.814178 INFO::Fitting model to feature number 17, Clostridium.sensu.stricto.1
#> 2024-05-29 18:02:11.818679 INFO::Fitting model to feature number 18, Colidextribacter
#> 2024-05-29 18:02:11.82313 INFO::Fitting model to feature number 19, Collinsella
#> 2024-05-29 18:02:11.827605 INFO::Fitting model to feature number 20, Coprobacillus
#> 2024-05-29 18:02:11.83212 INFO::Fitting model to feature number 21, Coprococcus
#> 2024-05-29 18:02:11.836597 INFO::Fitting model to feature number 22, DTU089
#> 2024-05-29 18:02:11.841038 INFO::Fitting model to feature number 23, Dialister
#> 2024-05-29 18:02:11.845565 INFO::Fitting model to feature number 24, Dorea
#> 2024-05-29 18:02:11.850054 INFO::Fitting model to feature number 25, Eggerthella
#> 2024-05-29 18:02:11.854522 INFO::Fitting model to feature number 26, Eisenbergiella
#> 2024-05-29 18:02:11.858963 INFO::Fitting model to feature number 27, Enterococcus
#> 2024-05-29 18:02:11.863423 INFO::Fitting model to feature number 28, Erysipelatoclostridium
#> 2024-05-29 18:02:11.867886 INFO::Fitting model to feature number 29, Erysipelotrichaceae.UCG.003
#> 2024-05-29 18:02:11.872348 INFO::Fitting model to feature number 30, Escherichia.Shigella
#> 2024-05-29 18:02:11.876827 INFO::Fitting model to feature number 31, Faecalibacterium
#> 2024-05-29 18:02:11.881296 INFO::Fitting model to feature number 32, Family.XIII.AD3011.group
#> 2024-05-29 18:02:11.885733 INFO::Fitting model to feature number 33, Flavonifractor
#> 2024-05-29 18:02:11.890171 INFO::Fitting model to feature number 34, Fusicatenibacter
#> 2024-05-29 18:02:11.894624 INFO::Fitting model to feature number 35, Fusobacterium
#> 2024-05-29 18:02:11.899062 INFO::Fitting model to feature number 36, Gemella
#> 2024-05-29 18:02:11.90352 INFO::Fitting model to feature number 37, Granulicatella
#> 2024-05-29 18:02:11.908012 INFO::Fitting model to feature number 38, Haemophilus
#> 2024-05-29 18:02:11.912463 INFO::Fitting model to feature number 39, Holdemania
#> 2024-05-29 18:02:11.916933 INFO::Fitting model to feature number 40, Hungatella
#> 2024-05-29 18:02:11.921422 INFO::Fitting model to feature number 41, Incertae.Sedis..EUPATH_0009256_Bacteria_Firmicutes_Clostridia_Oscillospirales_Ruminococcaceae_Incertae_Sedis.
#> 2024-05-29 18:02:11.925894 INFO::Fitting model to feature number 42, Intestinibacter
#> 2024-05-29 18:02:11.930362 INFO::Fitting model to feature number 43, Intestinimonas
#> 2024-05-29 18:02:11.934886 INFO::Fitting model to feature number 44, Klebsiella
#> 2024-05-29 18:02:11.939345 INFO::Fitting model to feature number 45, Lachnoclostridium
#> 2024-05-29 18:02:11.953494 INFO::Fitting model to feature number 46, Lachnospira
#> 2024-05-29 18:02:11.958221 INFO::Fitting model to feature number 47, Lachnospiraceae.FCS020.group
#> 2024-05-29 18:02:11.962802 INFO::Fitting model to feature number 48, Lachnospiraceae.ND3007.group
#> 2024-05-29 18:02:11.96734 INFO::Fitting model to feature number 49, Lachnospiraceae.NK4A136.group
#> 2024-05-29 18:02:11.971883 INFO::Fitting model to feature number 50, Lachnospiraceae.UCG.001
#> 2024-05-29 18:02:11.976421 INFO::Fitting model to feature number 51, Lachnospiraceae.UCG.004
#> 2024-05-29 18:02:11.980973 INFO::Fitting model to feature number 52, Lachnospiraceae.UCG.009
#> 2024-05-29 18:02:11.985563 INFO::Fitting model to feature number 53, Lachnospiraceae.UCG.010
#> 2024-05-29 18:02:11.990072 INFO::Fitting model to feature number 54, Lactobacillus
#> 2024-05-29 18:02:11.994635 INFO::Fitting model to feature number 55, Lactococcus
#> 2024-05-29 18:02:11.999248 INFO::Fitting model to feature number 56, Mannheimia
#> 2024-05-29 18:02:12.003996 INFO::Fitting model to feature number 57, Megasphaera
#> 2024-05-29 18:02:12.009041 INFO::Fitting model to feature number 58, Monoglobus
#> 2024-05-29 18:02:12.01379 INFO::Fitting model to feature number 59, Neisseria
#> 2024-05-29 18:02:12.018389 INFO::Fitting model to feature number 60, Odoribacter
#> 2024-05-29 18:02:12.022917 INFO::Fitting model to feature number 61, Oscillibacter
#> 2024-05-29 18:02:12.027505 INFO::Fitting model to feature number 62, Oscillospira
#> 2024-05-29 18:02:12.032041 INFO::Fitting model to feature number 63, Parabacteroides
#> 2024-05-29 18:02:12.036576 INFO::Fitting model to feature number 64, Parasutterella
#> 2024-05-29 18:02:12.041085 INFO::Fitting model to feature number 65, Prevotella
#> 2024-05-29 18:02:12.045629 INFO::Fitting model to feature number 66, Pseudomonas
#> 2024-05-29 18:02:12.05011 INFO::Fitting model to feature number 67, Romboutsia
#> 2024-05-29 18:02:12.054622 INFO::Fitting model to feature number 68, Roseburia
#> 2024-05-29 18:02:12.059153 INFO::Fitting model to feature number 69, Rothia
#> 2024-05-29 18:02:12.063628 INFO::Fitting model to feature number 70, Ruminococcus
#> 2024-05-29 18:02:12.068085 INFO::Fitting model to feature number 71, Sellimonas
#> 2024-05-29 18:02:12.072545 INFO::Fitting model to feature number 72, Solobacterium
#> 2024-05-29 18:02:12.077047 INFO::Fitting model to feature number 73, Staphylococcus
#> 2024-05-29 18:02:12.081558 INFO::Fitting model to feature number 74, Streptococcus
#> 2024-05-29 18:02:12.086076 INFO::Fitting model to feature number 75, Subdoligranulum
#> 2024-05-29 18:02:12.090592 INFO::Fitting model to feature number 76, Sutterella
#> 2024-05-29 18:02:12.095082 INFO::Fitting model to feature number 77, Terrisporobacter
#> 2024-05-29 18:02:12.099575 INFO::Fitting model to feature number 78, Turicibacter
#> 2024-05-29 18:02:12.104083 INFO::Fitting model to feature number 79, Tuzzerella
#> 2024-05-29 18:02:12.108549 INFO::Fitting model to feature number 80, Tyzzerella
#> 2024-05-29 18:02:12.112993 INFO::Fitting model to feature number 81, UBA1819
#> 2024-05-29 18:02:12.117599 INFO::Fitting model to feature number 82, UCG.002
#> 2024-05-29 18:02:12.122085 INFO::Fitting model to feature number 83, UCG.003
#> 2024-05-29 18:02:12.126568 INFO::Fitting model to feature number 84, UCG.005
#> 2024-05-29 18:02:12.131033 INFO::Fitting model to feature number 85, Veillonella
#> 2024-05-29 18:02:12.135566 INFO::Fitting model to feature number 86, X.Clostridium..innocuum.group
#> 2024-05-29 18:02:12.140044 INFO::Fitting model to feature number 87, X.Eubacterium..eligens.group
#> 2024-05-29 18:02:12.144504 INFO::Fitting model to feature number 88, X.Eubacterium..fissicatena.group
#> 2024-05-29 18:02:12.158477 INFO::Fitting model to feature number 89, X.Eubacterium..hallii.group
#> 2024-05-29 18:02:12.163185 INFO::Fitting model to feature number 90, X.Eubacterium..siraeum.group
#> 2024-05-29 18:02:12.167799 INFO::Fitting model to feature number 91, X.Eubacterium..ventriosum.group
#> 2024-05-29 18:02:12.172334 INFO::Fitting model to feature number 92, X.Eubacterium..xylanophilum.group
#> 2024-05-29 18:02:12.176863 INFO::Fitting model to feature number 93, X.Ruminococcus..gauvreauii.group
#> 2024-05-29 18:02:12.181423 INFO::Fitting model to feature number 94, X.Ruminococcus..gnavus.group
#> 2024-05-29 18:02:12.185961 INFO::Fitting model to feature number 95, X.Ruminococcus..torques.group
#> 2024-05-29 18:02:12.190502 INFO::Fitting model to feature number 96, unclassified.Chloroplast
#> 2024-05-29 18:02:12.195013 INFO::Fitting model to feature number 97, unclassified.Clostridia.UCG.014
#> 2024-05-29 18:02:12.199538 INFO::Fitting model to feature number 98, unclassified.Clostridia
#> 2024-05-29 18:02:12.204066 INFO::Fitting model to feature number 99, unclassified.Clostridia.vadinBB60.group
#> 2024-05-29 18:02:12.208536 INFO::Fitting model to feature number 100, unclassified.Gastranaerophilales
#> 2024-05-29 18:02:12.213044 INFO::Fitting model to feature number 101, unclassified.Lachnospiraceae
#> 2024-05-29 18:02:12.217538 INFO::Fitting model to feature number 102, unclassified.Mitochondria
#> 2024-05-29 18:02:12.221994 INFO::Fitting model to feature number 103, unclassified.Oscillospiraceae
#> 2024-05-29 18:02:12.226492 INFO::Fitting model to feature number 104, unclassified.Peptoniphilus
#> 2024-05-29 18:02:12.23099 INFO::Fitting model to feature number 105, unclassified.Rhodospirillales
#> 2024-05-29 18:02:12.235504 INFO::Fitting model to feature number 106, unclassified.Ruminococcaceae
#> 2024-05-29 18:02:12.240052 INFO::Fitting model to feature number 107, unclassified..Eubacterium..coprostanoligenes.group
#> 2024-05-29 18:02:12.261024 INFO::Counting total values for each feature
#> 2024-05-29 18:02:12.26713 INFO::Writing filtered data to file /tmp/RtmpMVCAZO/maaslin17af633dda21/features/filtered_data.tsv
#> 2024-05-29 18:02:12.480398 INFO::Writing filtered, normalized data to file /tmp/RtmpMVCAZO/maaslin17af633dda21/features/filtered_data_norm.tsv
#> 2024-05-29 18:02:12.696199 INFO::Writing filtered, normalized, transformed data to file /tmp/RtmpMVCAZO/maaslin17af633dda21/features/filtered_data_norm_transformed.tsv
#> 2024-05-29 18:02:13.052607 INFO::Writing residuals to file /tmp/RtmpMVCAZO/maaslin17af633dda21/fits/residuals.rds
#> 2024-05-29 18:02:13.123672 INFO::Writing fitted values to file /tmp/RtmpMVCAZO/maaslin17af633dda21/fits/fitted.rds
#> 2024-05-29 18:02:13.14927 INFO::Writing all results to file (ordered by increasing q-values): /tmp/RtmpMVCAZO/maaslin17af633dda21/all_results.tsv
#> 2024-05-29 18:02:13.151035 INFO::Writing the significant results (those which are less than or equal to the threshold of 0.250000 ) to file (ordered by increasing q-values): /tmp/RtmpMVCAZO/maaslin17af633dda21/significant_results.tsv