Maaslin2 from a Collection

MaAsLin2 finds associations between microbiome meta-omics features and complex metadata in population-scale epidemiological studies. The software includes multiple analysis methods (including support for multiple covariates and repeated measures) filtering, normalization, and transform options to customize analysis for your specific study.

Usage

Maaslin2(data, verbose = c(TRUE, FALSE), ...)

# S4 method for CollectionWithMetadata
Maaslin2(data, verbose = c(TRUE, FALSE), ...)

Arguments

data

a CollectionWithMetadata

verbose

boolean indicating if timed logging is desired

...

additional arguments to pass to Maaslin2::Maaslin2

Details

The MicrobiomeDB wrapper for Maaslin2 will ignore samples where the fixed_effects variable has a value of NA or the empty string ("").

Examples

maaslinOutput <- MicrobiomeDB::Maaslin2(
       data = getCollection(
  microbiomeData::DiabImmune, 
  '16S (V4) Genus (Relative taxonomic abundance analysis)'), 
       output = tempfile("maaslin"),
       #min_prevalence = 0,
       fixed_effects = 'delivery_mode',
       analysis_method = "LM", # default LM
       normalization = "TSS", # default TSS
       transform = "LOG", # default LOG
       plot_heatmap = FALSE,
       plot_scatter = FALSE)
#> [1] "Creating output folder"
#> [1] "Creating output feature tables folder"
#> [1] "Creating output fits folder"
#> 2024-06-26 14:27:29.682172 INFO::Writing function arguments to log file
#> 2024-06-26 14:27:29.696927 INFO::Verifying options selected are valid
#> 2024-06-26 14:27:29.724798 INFO::Determining format of input files
#> 2024-06-26 14:27:29.726582 INFO::Input format is data samples as rows and metadata samples as rows
#> 2024-06-26 14:27:29.741943 INFO::Formula for fixed effects: expr ~  delivery_mode
#> 2024-06-26 14:27:29.742946 INFO::Filter data based on min abundance and min prevalence
#> 2024-06-26 14:27:29.743443 INFO::Total samples in data: 3184
#> 2024-06-26 14:27:29.743913 INFO::Min samples required with min abundance for a feature not to be filtered: 318.400000
#> 2024-06-26 14:27:29.774512 INFO::Total filtered features: 564
#> 2024-06-26 14:27:29.77556 INFO::Filtered feature names from abundance and prevalence filtering: X28.4, Abiotrophia, Acetanaerobacterium, Acetobacter, Achromobacter, Acidaminococcus, Acidibacter, Acidiphilium, Acidipropionibacterium, Acidocella, Acidovorax, Acinetobacter, Actinobacillus, Actinobaculum, Actinotignum, Adlercreutzia, Aerococcus, Aeromonas, Afipia, Aggregatibacter, Alcaligenes, Alicycliphilus, Allisonella, Allobaculum, Alloprevotella, Allorhizobium.Neorhizobium.Pararhizobium.Rhizobium, Alloscardovia, Amaricoccus, Amnibacterium, Anaerobium, Anaerocolumna, Anaerofilum, Anaerofustis, Anaeroglobus, Anaeromyxobacter, Anaeroplasma, Anaerosporobacter, Anaerosporomusa, Anaerostignum, Anaerotruncus, Angelakisella, Angustibacter, Anoxybacillus, Aquabacterium, Aquamonas, Arcanobacterium, Arcticibacter, Asaccharospora, Asaia, Asinibacterium, Asteroleplasma, Asticcacaulis, Aureimonas, Azospira, Bacillus, Bacteriovorax, Bdellovibrio, Belnapia, Bergeyella, Blastomonas, Bordetella, Bosea, Brachybacterium, Bradyrhizobium, Brevibacterium, Brevundimonas, Brucella, Bulleidia, Burkholderia.Caballeronia.Paraburkholderia, Buttiauxella, Butyricimonas, CAG.56, CAG.873, CHKCI001, CHKCI002, CL500.29.marine.group, Caloribacterium, Candidatus.Arthromitus, Candidatus.Metachlamydia, Candidatus.Obscuribacter, Candidatus.Phytoplasma, Candidatus.Soleaferrea, Candidatus.Stoquefichus, Capnocytophaga, Caproiciproducens, Cardiobacterium, Carnobacterium, Caryophanon, Castellaniella, Catabacter, Catenibacillus, Catenibacterium, Catenisphaera, Catonella, Caulobacter, Cellulomonas, Cellulosilyticum, Cellvibrio, Centipeda, Cetobacterium, Christensenella, Chryseobacterium, Chryseomicrobium, Chungangia, Citrobacter, Clavibacter, Cloacibacillus, Cloacibacterium, Clostridium.sensu.stricto.12, Clostridium.sensu.stricto.13, Clostridium.sensu.stricto.15, Clostridium.sensu.stricto.18, Clostridium.sensu.stricto.2, Clostridium.sensu.stricto.3, Clostridium.sensu.stricto.4, Clostridium.sensu.stricto.6, Comamonas, Conexibacter, Conservatibacter, Coprobacter, Coriobacteriaceae.UCG.002, Corynebacterium, Cosenzaea, Criibacterium, Cryptobacterium, Curvibacter, Cutibacterium, Cytophaga, DNF00809, DSSD61, Dechloromonas, Defluviicoccus, Defluviitaleaceae.UCG.011, Delftia, Denitrobacterium, Dermabacter, Desemzia, Desulfovibrio, Devosia, Diaphorobacter, Dickeya, Dielma, Dietzia, Dinghuibacter, Diplosphaera, Dolosigranulum, Duganella, Dysgonomonas, Edaphobaculum, Eikenella, Empedobacter, Endobacter, Enhydrobacter, Enterobacter, Enterorhabdus, Enteroscipio, Epulopiscium, Erwinia, Erysipelotrichaceae.UCG.006, Eubacterium, Exiguobacterium, Extensimonas, F0058, F0332, FD2005, Facklamia, Faecalicoccus, Faecalitalea, Falsochrobactrum, Family.XIII.UCG.001, Fictibacillus, Flaviflexus, Flavihumibacter, Flavobacterium, Floricoccus, Fluviicola, Fournierella, Fretibacterium, Frisingicoccus, Fructobacillus, GCA.900066575, GCA.900066755, Galbibacter, Garciella, Gemmata, Geobacillus, Gluconobacter, Glutamicibacter, Gordonia, Gordonibacter, Gulbenkiania, Hafnia.Obesumbacterium, Halomonas, Haoranjiania, Harryflintia, Helicobacter, Hephaestia, Herbaspirillum, Herbinix, Holdemanella, Howardella, Hydrocarboniphaga, Hydrogenophilus, Hymenobacter, Iamia, Ignavigranum, Incertae.Sedis..EUPATH_0009256_Bacteria_Firmicutes_Clostridia_Lachnospirales_Lachnospiraceae_Incertae_Sedis., Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Erwiniaceae_Incertae_Sedis., Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Morganellaceae_Incertae_Sedis., Janthinobacterium, Jeotgalicoccus, Johnsonella, Kingella, Knoellia, Kocuria, Kosakonia, Lachnoanaerobaculum, Lachnobacterium, Lachnospiraceae.AC2044.group, Lachnospiraceae.FE2018.group, Lachnospiraceae.NC2004.group, Lachnospiraceae.UCG.002, Lachnospiraceae.UCG.003, Lachnospiraceae.UCG.008, Lachnospiraceae.XPB1014.group, Lachnotalea, Lactonifactor, Lautropia, Lawsonella, Legionella, Leptotrichia, Leuconostoc, Libanicoccus, Luteimonas, Luteolibacter, Macellibacteroides, Macrococcus, Mailhella, Marinilactibacillus, Marmoricola, Martelella, Marvinbryantia, Massilia, Megamonas, Merdibacter, Mesorhizobium, Methanobrevibacter, Methylibium, Methylobacillus, Methylobacterium.Methylorubrum, Methylorosula, Methylotenera, Micrococcus, Mitsuokella, Mobiluncus, Mogibacterium, Moheibacter, Mongoliitalea, Moraxella, Morganella, Moryella, Mucilaginibacter, Muribaculum, Murimonas, Mycobacterium, Mycoplasma, Myxococcus, NK4A214.group, Nakamurella, Nannocystis, Necropsobacter, Negativibacillus, Negativicoccus, Neochlamydia, Neomegalonema, Neorhizobium, Neoscardovia, Nesterenkonia, Nocardioides, Nocardiopsis, Novosphingobium, Oceanivirga, Oceanobacillus, Ochrobactrum, Olsenella, Oribacterium, Ottowia, Oxalobacter, Oxobacter, P3OB.42, Paenibacillus, Paeniclostridium, Paenisporosarcina, Paludicola, Papillibacter, Paraclostridium, Paracoccus, Paramesorhizobium, Paraprevotella, Parapusillimonas, Parascardovia, Parvibacter, Pasteurella, Paucibacter, Pediococcus, Pedobacter, Peptoclostridium, Peptococcus, Peptostreptococcus, Perlucidibaca, Phascolarctobacterium, Phaseolibacter, Phocea, Phoenicibacter, Phyllobacterium, Piscinibacter, Planifilum, Plesiomonas, Plot4.2H12, Pluralibacter, Polaromonas, Porphyromonas, Prevotellaceae.NK3B31.group, Prevotellaceae.UCG.001, Propionibacterium, Propionimicrobium, Prosthecobacter, Proteus, Providencia, Pseudacidovorax, Pseudaminobacter, Pseudocitrobacter, Pseudoflavonifractor, Pseudoglutamicibacter, Pseudoramibacter, Pseudoxanthomonas, Psychrobacter, Pygmaiobacter, Pyramidobacter, Qipengyuania, Ralstonia, Raoultella, Raoultibacter, Reyranella, Rhizorhapis, Rhodanobacter, Rhodobacter, Rhodococcus, Rikenella, Rikenellaceae.RC9.gut.group, Rivicola, Robinsoniella, Rodentibacter, Roseisolibacter, Rubellimicrobium, Rummeliibacillus, S5.A14a, SH.PL14, Saccharothrix, Salinimicrobium, Salinisphaera, Salipaludibacillus, Salmonella, Sanguibacteroides, Sarcina, Scardovia, Sedimentibacter, Segetibacter, Selenomonas, Senegalimassilia, Serratia, Shimwellia, Shuttleworthia, Slackia, Sneathia, Solibacillus, Sphingoaurantiacus, Sphingobacterium, Sphingobium, Sphingomonas, Sphingopyxis, Sporolactobacillus, Stenotrophomonas, Stomatobaculum, Streptobacillus, Succinatimonas, Succiniclasticum, Succinivibrio, Sulfurospirillum, Synergistes, TM7x, Tannerella, Tatumella, Tepidimonas, Tepidiphilus, Tetragenococcus, Thermus, Timonella, Tolumonas, Trabulsiella, Tropheryma, Trueperella, UC5.1.2E3, UCG.004, UCG.007, UCG.008, UCG.009, Ureibacillus, Varibaculum, Variovorax, Verticiella, Victivallis, Vulcaniibacterium, Weissella, Xanthomonas, Yersinia, Yokenella, Zoogloea, X.Bacteroides..pectinophilus.group, X.Eubacterium..brachy.group, X.Eubacterium..nodatum.group, X.Eubacterium..oxidoreducens.group, X.Eubacterium..ruminantium.group, unclassified.0319.6G20, unclassified.67.14, unclassified.A0839, unclassified.Acetobacteraceae, unclassified.Acidaminococcaceae, unclassified.Actinobacteria, unclassified.Actinomycetaceae, unclassified.Aeromonadales, unclassified.Anaerococcus, unclassified.Anaerosalibacter, unclassified.Anaerovoracaceae, unclassified.Armatimonadales, unclassified.Atopobiaceae, unclassified.Bacillaceae, unclassified.Bacillales, unclassified.Bacilli, unclassified.Bacteroidales, unclassified.Bacteroidia, unclassified.Barnesiellaceae, unclassified.Bifidobacteriaceae, unclassified.Burkholderiales, unclassified.Butyricicoccaceae, unclassified.Caloramatoraceae, unclassified.Carnobacteriaceae, unclassified.Caulobacteraceae, unclassified.Christensenellaceae, unclassified.Chroococcidiopsaceae, unclassified.Clostridiaceae, unclassified.Clostridiales, unclassified.Comamonadaceae, unclassified.Coriobacteriales.Incertae.Sedis..EUPATH_0009256_Bacteria_Actinobacteriota_Coriobacteriia_Coriobacteriales_Coriobacteriales_Incertae_Sedis_., unclassified.Coriobacteriales, unclassified.Corynebacteriaceae, unclassified.Desulfovibrionaceae, unclassified.Devosiaceae, unclassified.Diplorickettsiaceae, unclassified.Eggerthellaceae, unclassified.Enterobacterales, unclassified.Enterobacteriaceae, unclassified.Enterococcaceae, unclassified.Erwiniaceae, unclassified.Erysipelatoclostridiaceae, unclassified.Erysipelotrichaceae, unclassified.Erysipelotrichales, unclassified.Ethanoligenenaceae, unclassified.Eubacteriaceae, unclassified.Ezakiella, unclassified.Fastidiosipila, unclassified.Fenollaria, unclassified.Fimbriimonadaceae, unclassified.Finegoldia, unclassified.Firmicutes, unclassified.Flavobacteriaceae, unclassified.Flavobacteriales, unclassified.Fusobacteriaceae, unclassified.Gaiellales, unclassified.Gallicola, unclassified.Gammaproteobacteria, unclassified.Gemmataceae, unclassified.Hungateiclostridiaceae, unclassified.Hydrogenoanaerobacterium, unclassified.Izemoplasmatales, unclassified.JI49D030, unclassified.KD3.93, unclassified.Kapabacteriales, unclassified.Lachnospirales, unclassified.Lactobacillales, unclassified.Leuconostocaceae, unclassified.Lineage.IV, unclassified.MAT.CR.H4.C10, unclassified.Methylococcaceae, unclassified.Micavibrionales, unclassified.Microbacteriaceae, unclassified.Micrococcaceae, unclassified.Micrococcales, unclassified.Moraxellaceae, unclassified.Murdochiella, unclassified.Muribaculaceae, unclassified.NS9.marine.group, unclassified.Neisseriaceae, unclassified.OLB14, unclassified.Opitutales, unclassified.Oscillospirales, unclassified.Oxalobacteraceae, unclassified.P5D1.392, unclassified.PLTA13, unclassified.Paenibacillaceae, unclassified.Paracaedibacteraceae, unclassified.Parachlamydiaceae, unclassified.Parvimonas, unclassified.Pasteurellaceae, unclassified.Peptococcaceae, unclassified.Peptostreptococcaceae, unclassified.Peptostreptococcales.Tissierellales, unclassified.Planctomycetales, unclassified.Planococcaceae, unclassified.Prevotellaceae, unclassified.Promicromonosporaceae, unclassified.Propionibacteriaceae, unclassified.Proteobacteria, unclassified.Pseudomonadaceae, unclassified.Puniceicoccaceae, unclassified.RF39, unclassified.Rhizobiaceae, unclassified.Rhizobiales.Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Alphaproteobacteria_Rhizobiales_Rhizobiales_Incertae_Sedis_., unclassified.Rhizobiales, unclassified.Rhodobacteraceae, unclassified.Rhodocyclaceae, unclassified.Rickettsiales, unclassified.Rikenellaceae, unclassified.Ruminiclostridium, unclassified.ST.12K33, unclassified.Saccharimonadales, unclassified.Sandaracinaceae, unclassified.Selenomonadaceae, unclassified.Simkaniaceae, unclassified.Solimonadaceae, unclassified.Sphingobacteriales, unclassified.Sphingomonadaceae, unclassified.Staphylococcaceae, unclassified.Streptococcaceae, unclassified.Succinivibrionaceae, unclassified.Sutterellaceae, unclassified.TRA3.20, unclassified.Tannerellaceae, unclassified.Tepidimicrobium, unclassified.UCG.010, unclassified.UCG.011, unclassified.Veillonellaceae, unclassified.Veillonellales.Selenomonadales, unclassified.Vermiphilaceae, unclassified.Verrucomicrobiales, unclassified.Verrucomicrobiota, unclassified.Victivallaceae, unclassified.W5053, unclassified.WPS.2, unclassified.Weeksellaceae, unclassified.Xanthobacteraceae, unclassified.Yersiniaceae, unclassified..Clostridium..methylpentosum.group, unclassified.vadinBE97
#> 2024-06-26 14:27:29.78355 INFO::Total filtered features with variance filtering: 0
#> 2024-06-26 14:27:29.78413 INFO::Filtered feature names from variance filtering:
#> 2024-06-26 14:27:29.784617 INFO::Running selected normalization method: TSS
#> 2024-06-26 14:27:29.846918 INFO::Applying z-score to standardize continuous metadata
#> 2024-06-26 14:27:29.855053 INFO::Running selected transform method: LOG
#> 2024-06-26 14:27:29.929403 INFO::Running selected analysis method: LM
#> 2024-06-26 14:27:29.935358 INFO::Fitting model to feature number 1, Actinomyces
#> 2024-06-26 14:27:30.68271 INFO::Fitting model to feature number 2, Agathobacter
#> 2024-06-26 14:27:30.688327 INFO::Fitting model to feature number 3, Akkermansia
#> 2024-06-26 14:27:30.693074 INFO::Fitting model to feature number 4, Alistipes
#> 2024-06-26 14:27:30.697743 INFO::Fitting model to feature number 5, Anaerostipes
#> 2024-06-26 14:27:30.702388 INFO::Fitting model to feature number 6, Atopobium
#> 2024-06-26 14:27:30.706916 INFO::Fitting model to feature number 7, Bacteroides
#> 2024-06-26 14:27:30.711432 INFO::Fitting model to feature number 8, Barnesiella
#> 2024-06-26 14:27:30.715945 INFO::Fitting model to feature number 9, Bifidobacterium
#> 2024-06-26 14:27:30.72057 INFO::Fitting model to feature number 10, Bilophila
#> 2024-06-26 14:27:30.725066 INFO::Fitting model to feature number 11, Blautia
#> 2024-06-26 14:27:30.729595 INFO::Fitting model to feature number 12, Butyricicoccus
#> 2024-06-26 14:27:30.734111 INFO::Fitting model to feature number 13, CAG.352
#> 2024-06-26 14:27:30.738599 INFO::Fitting model to feature number 14, Campylobacter
#> 2024-06-26 14:27:30.743064 INFO::Fitting model to feature number 15, Christensenellaceae.R.7.group
#> 2024-06-26 14:27:30.74753 INFO::Fitting model to feature number 16, Clostridioides
#> 2024-06-26 14:27:30.75198 INFO::Fitting model to feature number 17, Clostridium.sensu.stricto.1
#> 2024-06-26 14:27:30.756477 INFO::Fitting model to feature number 18, Colidextribacter
#> 2024-06-26 14:27:30.760958 INFO::Fitting model to feature number 19, Collinsella
#> 2024-06-26 14:27:30.765425 INFO::Fitting model to feature number 20, Coprobacillus
#> 2024-06-26 14:27:30.769905 INFO::Fitting model to feature number 21, Coprococcus
#> 2024-06-26 14:27:30.774372 INFO::Fitting model to feature number 22, DTU089
#> 2024-06-26 14:27:30.778799 INFO::Fitting model to feature number 23, Dialister
#> 2024-06-26 14:27:30.783242 INFO::Fitting model to feature number 24, Dorea
#> 2024-06-26 14:27:30.787678 INFO::Fitting model to feature number 25, Eggerthella
#> 2024-06-26 14:27:30.792142 INFO::Fitting model to feature number 26, Eisenbergiella
#> 2024-06-26 14:27:30.796595 INFO::Fitting model to feature number 27, Enterococcus
#> 2024-06-26 14:27:30.801078 INFO::Fitting model to feature number 28, Erysipelatoclostridium
#> 2024-06-26 14:27:30.805504 INFO::Fitting model to feature number 29, Erysipelotrichaceae.UCG.003
#> 2024-06-26 14:27:30.809915 INFO::Fitting model to feature number 30, Escherichia.Shigella
#> 2024-06-26 14:27:30.814348 INFO::Fitting model to feature number 31, Faecalibacterium
#> 2024-06-26 14:27:30.818783 INFO::Fitting model to feature number 32, Family.XIII.AD3011.group
#> 2024-06-26 14:27:30.823269 INFO::Fitting model to feature number 33, Flavonifractor
#> 2024-06-26 14:27:30.827777 INFO::Fitting model to feature number 34, Fusicatenibacter
#> 2024-06-26 14:27:30.832203 INFO::Fitting model to feature number 35, Fusobacterium
#> 2024-06-26 14:27:30.836623 INFO::Fitting model to feature number 36, Gemella
#> 2024-06-26 14:27:30.841048 INFO::Fitting model to feature number 37, Granulicatella
#> 2024-06-26 14:27:30.845501 INFO::Fitting model to feature number 38, Haemophilus
#> 2024-06-26 14:27:30.849955 INFO::Fitting model to feature number 39, Holdemania
#> 2024-06-26 14:27:30.854382 INFO::Fitting model to feature number 40, Hungatella
#> 2024-06-26 14:27:30.858769 INFO::Fitting model to feature number 41, Incertae.Sedis..EUPATH_0009256_Bacteria_Firmicutes_Clostridia_Oscillospirales_Ruminococcaceae_Incertae_Sedis.
#> 2024-06-26 14:27:30.863179 INFO::Fitting model to feature number 42, Intestinibacter
#> 2024-06-26 14:27:30.867609 INFO::Fitting model to feature number 43, Intestinimonas
#> 2024-06-26 14:27:30.872024 INFO::Fitting model to feature number 44, Klebsiella
#> 2024-06-26 14:27:30.876467 INFO::Fitting model to feature number 45, Lachnoclostridium
#> 2024-06-26 14:27:30.880857 INFO::Fitting model to feature number 46, Lachnospira
#> 2024-06-26 14:27:30.886671 INFO::Fitting model to feature number 47, Lachnospiraceae.FCS020.group
#> 2024-06-26 14:27:30.892635 INFO::Fitting model to feature number 48, Lachnospiraceae.ND3007.group
#> 2024-06-26 14:27:30.898595 INFO::Fitting model to feature number 49, Lachnospiraceae.NK4A136.group
#> 2024-06-26 14:27:30.904554 INFO::Fitting model to feature number 50, Lachnospiraceae.UCG.001
#> 2024-06-26 14:27:30.910494 INFO::Fitting model to feature number 51, Lachnospiraceae.UCG.004
#> 2024-06-26 14:27:30.916446 INFO::Fitting model to feature number 52, Lachnospiraceae.UCG.009
#> 2024-06-26 14:27:30.922379 INFO::Fitting model to feature number 53, Lachnospiraceae.UCG.010
#> 2024-06-26 14:27:30.928372 INFO::Fitting model to feature number 54, Lactobacillus
#> 2024-06-26 14:27:30.93433 INFO::Fitting model to feature number 55, Lactococcus
#> 2024-06-26 14:27:30.940294 INFO::Fitting model to feature number 56, Mannheimia
#> 2024-06-26 14:27:30.946261 INFO::Fitting model to feature number 57, Megasphaera
#> 2024-06-26 14:27:30.952261 INFO::Fitting model to feature number 58, Monoglobus
#> 2024-06-26 14:27:30.958254 INFO::Fitting model to feature number 59, Neisseria
#> 2024-06-26 14:27:30.964304 INFO::Fitting model to feature number 60, Odoribacter
#> 2024-06-26 14:27:30.970278 INFO::Fitting model to feature number 61, Oscillibacter
#> 2024-06-26 14:27:30.97627 INFO::Fitting model to feature number 62, Oscillospira
#> 2024-06-26 14:27:30.982278 INFO::Fitting model to feature number 63, Parabacteroides
#> 2024-06-26 14:27:30.988256 INFO::Fitting model to feature number 64, Parasutterella
#> 2024-06-26 14:27:30.99433 INFO::Fitting model to feature number 65, Prevotella
#> 2024-06-26 14:27:31.00039 INFO::Fitting model to feature number 66, Pseudomonas
#> 2024-06-26 14:27:31.006396 INFO::Fitting model to feature number 67, Romboutsia
#> 2024-06-26 14:27:31.012378 INFO::Fitting model to feature number 68, Roseburia
#> 2024-06-26 14:27:31.018394 INFO::Fitting model to feature number 69, Rothia
#> 2024-06-26 14:27:31.024542 INFO::Fitting model to feature number 70, Ruminococcus
#> 2024-06-26 14:27:31.030638 INFO::Fitting model to feature number 71, Sellimonas
#> 2024-06-26 14:27:31.03677 INFO::Fitting model to feature number 72, Solobacterium
#> 2024-06-26 14:27:31.042808 INFO::Fitting model to feature number 73, Staphylococcus
#> 2024-06-26 14:27:31.04881 INFO::Fitting model to feature number 74, Streptococcus
#> 2024-06-26 14:27:31.055039 INFO::Fitting model to feature number 75, Subdoligranulum
#> 2024-06-26 14:27:31.061346 INFO::Fitting model to feature number 76, Sutterella
#> 2024-06-26 14:27:31.067462 INFO::Fitting model to feature number 77, Terrisporobacter
#> 2024-06-26 14:27:31.073571 INFO::Fitting model to feature number 78, Turicibacter
#> 2024-06-26 14:27:31.079634 INFO::Fitting model to feature number 79, Tuzzerella
#> 2024-06-26 14:27:31.085753 INFO::Fitting model to feature number 80, Tyzzerella
#> 2024-06-26 14:27:31.091835 INFO::Fitting model to feature number 81, UBA1819
#> 2024-06-26 14:27:31.097953 INFO::Fitting model to feature number 82, UCG.002
#> 2024-06-26 14:27:31.104068 INFO::Fitting model to feature number 83, UCG.003
#> 2024-06-26 14:27:31.110127 INFO::Fitting model to feature number 84, UCG.005
#> 2024-06-26 14:27:31.116212 INFO::Fitting model to feature number 85, Veillonella
#> 2024-06-26 14:27:31.12236 INFO::Fitting model to feature number 86, X.Clostridium..innocuum.group
#> 2024-06-26 14:27:31.128447 INFO::Fitting model to feature number 87, X.Eubacterium..eligens.group
#> 2024-06-26 14:27:31.142886 INFO::Fitting model to feature number 88, X.Eubacterium..fissicatena.group
#> 2024-06-26 14:27:31.14741 INFO::Fitting model to feature number 89, X.Eubacterium..hallii.group
#> 2024-06-26 14:27:31.151888 INFO::Fitting model to feature number 90, X.Eubacterium..siraeum.group
#> 2024-06-26 14:27:31.156343 INFO::Fitting model to feature number 91, X.Eubacterium..ventriosum.group
#> 2024-06-26 14:27:31.160805 INFO::Fitting model to feature number 92, X.Eubacterium..xylanophilum.group
#> 2024-06-26 14:27:31.165271 INFO::Fitting model to feature number 93, X.Ruminococcus..gauvreauii.group
#> 2024-06-26 14:27:31.169882 INFO::Fitting model to feature number 94, X.Ruminococcus..gnavus.group
#> 2024-06-26 14:27:31.174381 INFO::Fitting model to feature number 95, X.Ruminococcus..torques.group
#> 2024-06-26 14:27:31.178882 INFO::Fitting model to feature number 96, unclassified.Chloroplast
#> 2024-06-26 14:27:31.183355 INFO::Fitting model to feature number 97, unclassified.Clostridia.UCG.014
#> 2024-06-26 14:27:31.187812 INFO::Fitting model to feature number 98, unclassified.Clostridia
#> 2024-06-26 14:27:31.192285 INFO::Fitting model to feature number 99, unclassified.Clostridia.vadinBB60.group
#> 2024-06-26 14:27:31.196805 INFO::Fitting model to feature number 100, unclassified.Gastranaerophilales
#> 2024-06-26 14:27:31.201344 INFO::Fitting model to feature number 101, unclassified.Lachnospiraceae
#> 2024-06-26 14:27:31.205845 INFO::Fitting model to feature number 102, unclassified.Mitochondria
#> 2024-06-26 14:27:31.210337 INFO::Fitting model to feature number 103, unclassified.Oscillospiraceae
#> 2024-06-26 14:27:31.214799 INFO::Fitting model to feature number 104, unclassified.Peptoniphilus
#> 2024-06-26 14:27:31.219235 INFO::Fitting model to feature number 105, unclassified.Rhodospirillales
#> 2024-06-26 14:27:31.223686 INFO::Fitting model to feature number 106, unclassified.Ruminococcaceae
#> 2024-06-26 14:27:31.228182 INFO::Fitting model to feature number 107, unclassified..Eubacterium..coprostanoligenes.group
#> 2024-06-26 14:27:31.247113 INFO::Counting total values for each feature
#> 2024-06-26 14:27:31.252876 INFO::Writing filtered data to file /tmp/Rtmp1DvrdC/maaslin17d61689e91c/features/filtered_data.tsv
#> 2024-06-26 14:27:31.470656 INFO::Writing filtered, normalized data to file /tmp/Rtmp1DvrdC/maaslin17d61689e91c/features/filtered_data_norm.tsv
#> 2024-06-26 14:27:31.691205 INFO::Writing filtered, normalized, transformed data to file /tmp/Rtmp1DvrdC/maaslin17d61689e91c/features/filtered_data_norm_transformed.tsv
#> 2024-06-26 14:27:32.052359 INFO::Writing residuals to file /tmp/Rtmp1DvrdC/maaslin17d61689e91c/fits/residuals.rds
#> 2024-06-26 14:27:32.123329 INFO::Writing fitted values to file /tmp/Rtmp1DvrdC/maaslin17d61689e91c/fits/fitted.rds
#> 2024-06-26 14:27:32.147751 INFO::Writing all results to file (ordered by increasing q-values): /tmp/Rtmp1DvrdC/maaslin17d61689e91c/all_results.tsv
#> 2024-06-26 14:27:32.149597 INFO::Writing the significant results (those which are less than or equal to the threshold of 0.250000 ) to file (ordered by increasing q-values): /tmp/Rtmp1DvrdC/maaslin17d61689e91c/significant_results.tsv