Maaslin2 from a Collection

MaAsLin2 finds associations between microbiome meta-omics features and complex metadata in population-scale epidemiological studies. The software includes multiple analysis methods (including support for multiple covariates and repeated measures) filtering, normalization, and transform options to customize analysis for your specific study.

Usage

Maaslin2(data, verbose = c(TRUE, FALSE), ...)

# S4 method for CollectionWithMetadata
Maaslin2(data, verbose = c(TRUE, FALSE), ...)

Arguments

data

a CollectionWithMetadata

verbose

boolean indicating if timed logging is desired

...

additional arguments to pass to Maaslin2::Maaslin2

Details

The MicrobiomeDB wrapper for Maaslin2 will ignore samples where the fixed_effects variable has a value of NA or the empty string ("").

Examples

maaslinOutput <- MicrobiomeDB::Maaslin2(
       data = getCollection(microbiomeData::DiabImmune, '16S (V4) Genus'), 
       output = tempfile("maaslin"),
       #min_prevalence = 0,
       fixed_effects = 'delivery_mode',
       analysis_method = "LM", # default LM
       normalization = "TSS", # default TSS
       transform = "LOG", # default LOG
       plot_heatmap = FALSE,
       plot_scatter = FALSE)
#> [1] "Creating output folder"
#> [1] "Creating output feature tables folder"
#> [1] "Creating output fits folder"
#> 2024-04-30 19:50:16.090187 INFO::Writing function arguments to log file
#> 2024-04-30 19:50:16.104682 INFO::Verifying options selected are valid
#> 2024-04-30 19:50:16.132489 INFO::Determining format of input files
#> 2024-04-30 19:50:16.134112 INFO::Input format is data samples as rows and metadata samples as rows
#> 2024-04-30 19:50:16.148365 INFO::Formula for fixed effects: expr ~  delivery_mode
#> 2024-04-30 19:50:16.149421 INFO::Filter data based on min abundance and min prevalence
#> 2024-04-30 19:50:16.149903 INFO::Total samples in data: 3184
#> 2024-04-30 19:50:16.150351 INFO::Min samples required with min abundance for a feature not to be filtered: 318.400000
#> 2024-04-30 19:50:16.169918 INFO::Total filtered features: 563
#> 2024-04-30 19:50:16.170969 INFO::Filtered feature names from abundance and prevalence filtering: X28.4, Abiotrophia, Acetanaerobacterium, Acetobacter, Achromobacter, Acidaminococcus, Acidibacter, Acidiphilium, Acidipropionibacterium, Acidocella, Acidovorax, Acinetobacter, Actinobacillus, Actinobaculum, Actinotignum, Adlercreutzia, Aerococcus, Aeromonas, Afipia, Aggregatibacter, Alcaligenes, Alicycliphilus, Allisonella, Allobaculum, Alloprevotella, Allorhizobium.Neorhizobium.Pararhizobium.Rhizobium, Alloscardovia, Amaricoccus, Amnibacterium, Anaerobium, Anaerocolumna, Anaerofilum, Anaerofustis, Anaeroglobus, Anaeromyxobacter, Anaeroplasma, Anaerosporobacter, Anaerosporomusa, Anaerostignum, Anaerotruncus, Angelakisella, Angustibacter, Anoxybacillus, Aquabacterium, Aquamonas, Arcanobacterium, Arcticibacter, Asaccharospora, Asaia, Asinibacterium, Asteroleplasma, Asticcacaulis, Aureimonas, Azospira, Bacillus, Bacteriovorax, Bdellovibrio, Belnapia, Bergeyella, Blastomonas, Bordetella, Bosea, Brachybacterium, Bradyrhizobium, Brevibacterium, Brevundimonas, Brucella, Bulleidia, Burkholderia.Caballeronia.Paraburkholderia, Buttiauxella, Butyricimonas, CAG.56, CAG.873, CHKCI001, CHKCI002, CL500.29.marine.group, Caloribacterium, Candidatus.Arthromitus, Candidatus.Metachlamydia, Candidatus.Obscuribacter, Candidatus.Phytoplasma, Candidatus.Soleaferrea, Candidatus.Stoquefichus, Capnocytophaga, Caproiciproducens, Cardiobacterium, Carnobacterium, Caryophanon, Castellaniella, Catabacter, Catenibacillus, Catenibacterium, Catenisphaera, Catonella, Caulobacter, Cellulomonas, Cellulosilyticum, Cellvibrio, Centipeda, Cetobacterium, Christensenella, Chryseobacterium, Chryseomicrobium, Chungangia, Citrobacter, Clavibacter, Cloacibacillus, Cloacibacterium, Clostridium.sensu.stricto.12, Clostridium.sensu.stricto.13, Clostridium.sensu.stricto.15, Clostridium.sensu.stricto.18, Clostridium.sensu.stricto.2, Clostridium.sensu.stricto.3, Clostridium.sensu.stricto.4, Clostridium.sensu.stricto.6, Comamonas, Conexibacter, Conservatibacter, Coprobacter, Coriobacteriaceae.UCG.002, Corynebacterium, Cosenzaea, Criibacterium, Cryptobacterium, Curvibacter, Cutibacterium, Cytophaga, DNF00809, DSSD61, Dechloromonas, Defluviicoccus, Defluviitaleaceae.UCG.011, Delftia, Denitrobacterium, Dermabacter, Desemzia, Desulfovibrio, Devosia, Diaphorobacter, Dickeya, Dielma, Dietzia, Dinghuibacter, Diplosphaera, Dolosigranulum, Duganella, Dysgonomonas, Edaphobaculum, Eikenella, Empedobacter, Endobacter, Enhydrobacter, Enterobacter, Enterorhabdus, Enteroscipio, Epulopiscium, Erwinia, Erysipelotrichaceae.UCG.006, Eubacterium, Exiguobacterium, Extensimonas, F0058, F0332, FD2005, Facklamia, Faecalicoccus, Faecalitalea, Falsochrobactrum, Family.XIII.UCG.001, Fictibacillus, Flaviflexus, Flavihumibacter, Flavobacterium, Floricoccus, Fluviicola, Fournierella, Fretibacterium, Frisingicoccus, Fructobacillus, GCA.900066575, GCA.900066755, Galbibacter, Garciella, Gemmata, Geobacillus, Gluconobacter, Glutamicibacter, Gordonia, Gordonibacter, Gulbenkiania, Hafnia.Obesumbacterium, Halomonas, Haoranjiania, Harryflintia, Helicobacter, Hephaestia, Herbaspirillum, Herbinix, Holdemanella, Howardella, Hydrocarboniphaga, Hydrogenophilus, Hymenobacter, Iamia, Ignavigranum, Incertae.Sedis..EUPATH_0009256_Bacteria_Firmicutes_Clostridia_Lachnospirales_Lachnospiraceae_Incertae_Sedis., Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Erwiniaceae_Incertae_Sedis., Incertae.Sedis..EUPATH_0009256_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Morganellaceae_Incertae_Sedis., Janthinobacterium, Jeotgalicoccus, Johnsonella, Kingella, Knoellia, Kocuria, Kosakonia, Lachnoanaerobaculum, Lachnobacterium, Lachnospiraceae.AC2044.group, Lachnospiraceae.FE2018.group, Lachnospiraceae.NC2004.group, Lachnospiraceae.UCG.002, Lachnospiraceae.UCG.003, Lachnospiraceae.UCG.008, Lachnospiraceae.XPB1014.group, Lachnotalea, Lactonifactor, Lautropia, Lawsonella, Legionella, Leptotrichia, Leuconostoc, Libanicoccus, Luteimonas, Luteolibacter, Macellibacteroides, Macrococcus, Mailhella, Marinilactibacillus, Marmoricola, Martelella, Marvinbryantia, Massilia, Megamonas, Merdibacter, Mesorhizobium, Methanobrevibacter, Methylibium, Methylobacillus, Methylobacterium.Methylorubrum, Methylorosula, Methylotenera, Micrococcus, Mitsuokella, Mobiluncus, Mogibacterium, Moheibacter, Mongoliitalea, Moraxella, Morganella, Moryella, Mucilaginibacter, Muribaculum, Murimonas, Mycobacterium, Mycoplasma, Myxococcus, NK4A214.group, Nakamurella, Nannocystis, Necropsobacter, Negativibacillus, Negativicoccus, Neochlamydia, Neomegalonema, Neorhizobium, Neoscardovia, Nesterenkonia, Nocardioides, Nocardiopsis, Novosphingobium, Oceanivirga, Oceanobacillus, Ochrobactrum, Olsenella, Oribacterium, Ottowia, Oxalobacter, Oxobacter, P3OB.42, Paenibacillus, Paeniclostridium, Paenisporosarcina, Paludicola, Papillibacter, Paraclostridium, Paracoccus, Paramesorhizobium, Paraprevotella, Parapusillimonas, Parascardovia, Parvibacter, Pasteurella, Paucibacter, Pediococcus, Pedobacter, Peptoclostridium, Peptococcus, Peptostreptococcus, Perlucidibaca, Phascolarctobacterium, Phaseolibacter, Phocea, Phoenicibacter, Phyllobacterium, Piscinibacter, Planifilum, Plesiomonas, Plot4.2H12, Pluralibacter, Polaromonas, Porphyromonas, Prevotellaceae.NK3B31.group, Prevotellaceae.UCG.001, Propionibacterium, Prosthecobacter, Proteus, Providencia, Pseudacidovorax, Pseudaminobacter, Pseudocitrobacter, Pseudoflavonifractor, Pseudoglutamicibacter, Pseudoramibacter, Pseudoxanthomonas, Psychrobacter, Pygmaiobacter, Pyramidobacter, Qipengyuania, Ralstonia, Raoultella, Raoultibacter, Reyranella, Rhizorhapis, Rhodanobacter, Rhodobacter, Rhodococcus, Rikenella, Rikenellaceae.RC9.gut.group, Rivicola, Robinsoniella, Rodentibacter, Roseisolibacter, Rubellimicrobium, Rummeliibacillus, S5.A14a, SH.PL14, Saccharothrix, Salinimicrobium, Salinisphaera, Salipaludibacillus, Salmonella, Sanguibacteroides, Sarcina, Scardovia, Sedimentibacter, Segetibacter, Selenomonas, Senegalimassilia, Serratia, Shimwellia, Shuttleworthia, Slackia, Sneathia, Solibacillus, Sphingoaurantiacus, Sphingobacterium, Sphingobium, Sphingomonas, Sphingopyxis, Sporolactobacillus, Stenotrophomonas, Stomatobaculum, Streptobacillus, Succinatimonas, Succiniclasticum, Succinivibrio, Sulfurospirillum, Synergistes, TM7x, Tannerella, Tatumella, Tepidimonas, Tepidiphilus, Tetragenococcus, Thermus, Timonella, Tolumonas, Trabulsiella, Tropheryma, Trueperella, UC5.1.2E3, UCG.004, UCG.007, UCG.008, UCG.009, Ureibacillus, Varibaculum, Variovorax, Verticiella, Victivallis, Vulcaniibacterium, Weissella, Xanthomonas, Yersinia, Yokenella, Zoogloea, X.Bacteroides..pectinophilus.group, X.Eubacterium..brachy.group, X.Eubacterium..nodatum.group, X.Eubacterium..oxidoreducens.group, X.Eubacterium..ruminantium.group, unclassified.0319.6G20, unclassified.67.14, unclassified.A0839, unclassified.Acetobacteraceae, unclassified.Acidaminococcaceae, unclassified.Actinobacteria, unclassified.Actinomycetaceae, unclassified.Aeromonadales, unclassified.Anaerococcus, unclassified.Anaerosalibacter, unclassified.Anaerovoracaceae, unclassified.Armatimonadales, unclassified.Atopobiaceae, unclassified.Bacillaceae, unclassified.Bacillales, unclassified.Bacilli, unclassified.Bacteroidales, unclassified.Bacteroidia, unclassified.Barnesiellaceae, unclassified.Bifidobacteriaceae, unclassified.Burkholderiales, unclassified.Butyricicoccaceae, unclassified.Caloramatoraceae, unclassified.Carnobacteriaceae, unclassified.Caulobacteraceae, unclassified.Christensenellaceae, unclassified.Chroococcidiopsaceae, unclassified.Clostridiaceae, unclassified.Clostridiales, unclassified.Comamonadaceae, unclassified.Coriobacteriales.Incertae.Sedis, unclassified.Coriobacteriales, unclassified.Corynebacteriaceae, unclassified.Desulfovibrionaceae, unclassified.Devosiaceae, unclassified.Diplorickettsiaceae, unclassified.Eggerthellaceae, unclassified.Enterobacterales, unclassified.Enterobacteriaceae, unclassified.Enterococcaceae, unclassified.Erwiniaceae, unclassified.Erysipelatoclostridiaceae, unclassified.Erysipelotrichaceae, unclassified.Erysipelotrichales, unclassified.Ethanoligenenaceae, unclassified.Eubacteriaceae, unclassified.Ezakiella, unclassified.Fastidiosipila, unclassified.Fenollaria, unclassified.Fimbriimonadaceae, unclassified.Finegoldia, unclassified.Firmicutes, unclassified.Flavobacteriaceae, unclassified.Flavobacteriales, unclassified.Fusobacteriaceae, unclassified.Gaiellales, unclassified.Gallicola, unclassified.Gammaproteobacteria, unclassified.Gemmataceae, unclassified.Hungateiclostridiaceae, unclassified.Hydrogenoanaerobacterium, unclassified.Izemoplasmatales, unclassified.JI49D030, unclassified.KD3.93, unclassified.Kapabacteriales, unclassified.Lachnospirales, unclassified.Lactobacillales, unclassified.Leuconostocaceae, unclassified.Lineage.IV, unclassified.MAT.CR.H4.C10, unclassified.Methylococcaceae, unclassified.Micavibrionales, unclassified.Microbacteriaceae, unclassified.Micrococcaceae, unclassified.Micrococcales, unclassified.Moraxellaceae, unclassified.Murdochiella, unclassified.Muribaculaceae, unclassified.NS9.marine.group, unclassified.Neisseriaceae, unclassified.OLB14, unclassified.Opitutales, unclassified.Oscillospirales, unclassified.Oxalobacteraceae, unclassified.P5D1.392, unclassified.PLTA13, unclassified.Paenibacillaceae, unclassified.Paracaedibacteraceae, unclassified.Parachlamydiaceae, unclassified.Parvimonas, unclassified.Pasteurellaceae, unclassified.Peptococcaceae, unclassified.Peptostreptococcaceae, unclassified.Peptostreptococcales.Tissierellales, unclassified.Planctomycetales, unclassified.Planococcaceae, unclassified.Prevotellaceae, unclassified.Promicromonosporaceae, unclassified.Propionibacteriaceae, unclassified.Proteobacteria, unclassified.Pseudomonadaceae, unclassified.Puniceicoccaceae, unclassified.RF39, unclassified.Rhizobiaceae, unclassified.Rhizobiales.Incertae.Sedis, unclassified.Rhizobiales, unclassified.Rhodobacteraceae, unclassified.Rhodocyclaceae, unclassified.Rickettsiales, unclassified.Rikenellaceae, unclassified.Ruminiclostridium, unclassified.ST.12K33, unclassified.Saccharimonadales, unclassified.Sandaracinaceae, unclassified.Selenomonadaceae, unclassified.Simkaniaceae, unclassified.Solimonadaceae, unclassified.Sphingobacteriales, unclassified.Sphingomonadaceae, unclassified.Staphylococcaceae, unclassified.Streptococcaceae, unclassified.Succinivibrionaceae, unclassified.Sutterellaceae, unclassified.TRA3.20, unclassified.Tannerellaceae, unclassified.Tepidimicrobium, unclassified.UCG.010, unclassified.UCG.011, unclassified.Veillonellaceae, unclassified.Veillonellales.Selenomonadales, unclassified.Vermiphilaceae, unclassified.Verrucomicrobiales, unclassified.Verrucomicrobiota, unclassified.Victivallaceae, unclassified.W5053, unclassified.WPS.2, unclassified.Weeksellaceae, unclassified.Xanthobacteraceae, unclassified.Yersiniaceae, unclassified..Clostridium..methylpentosum.group, unclassified.vadinBE97
#> 2024-04-30 19:50:16.188831 INFO::Total filtered features with variance filtering: 0
#> 2024-04-30 19:50:16.189493 INFO::Filtered feature names from variance filtering:
#> 2024-04-30 19:50:16.189957 INFO::Running selected normalization method: TSS
#> 2024-04-30 19:50:16.238369 INFO::Applying z-score to standardize continuous metadata
#> 2024-04-30 19:50:16.246374 INFO::Running selected transform method: LOG
#> 2024-04-30 19:50:16.30883 INFO::Running selected analysis method: LM
#> 2024-04-30 19:50:16.314819 INFO::Fitting model to feature number 1, Actinomyces
#> 2024-04-30 19:50:16.323025 INFO::Fitting model to feature number 2, Agathobacter
#> 2024-04-30 19:50:16.327253 INFO::Fitting model to feature number 3, Akkermansia
#> 2024-04-30 19:50:16.331437 INFO::Fitting model to feature number 4, Alistipes
#> 2024-04-30 19:50:16.356302 INFO::Fitting model to feature number 5, Anaerostipes
#> 2024-04-30 19:50:16.360694 INFO::Fitting model to feature number 6, Atopobium
#> 2024-04-30 19:50:16.364856 INFO::Fitting model to feature number 7, Bacteroides
#> 2024-04-30 19:50:16.369028 INFO::Fitting model to feature number 8, Barnesiella
#> 2024-04-30 19:50:16.373125 INFO::Fitting model to feature number 9, Bifidobacterium
#> 2024-04-30 19:50:16.377244 INFO::Fitting model to feature number 10, Bilophila
#> 2024-04-30 19:50:16.381385 INFO::Fitting model to feature number 11, Blautia
#> 2024-04-30 19:50:16.385487 INFO::Fitting model to feature number 12, Butyricicoccus
#> 2024-04-30 19:50:16.389612 INFO::Fitting model to feature number 13, CAG.352
#> 2024-04-30 19:50:16.393692 INFO::Fitting model to feature number 14, Campylobacter
#> 2024-04-30 19:50:16.397785 INFO::Fitting model to feature number 15, Christensenellaceae.R.7.group
#> 2024-04-30 19:50:16.401874 INFO::Fitting model to feature number 16, Clostridioides
#> 2024-04-30 19:50:16.406067 INFO::Fitting model to feature number 17, Clostridium.sensu.stricto.1
#> 2024-04-30 19:50:16.410216 INFO::Fitting model to feature number 18, Colidextribacter
#> 2024-04-30 19:50:16.414338 INFO::Fitting model to feature number 19, Collinsella
#> 2024-04-30 19:50:16.418494 INFO::Fitting model to feature number 20, Coprobacillus
#> 2024-04-30 19:50:16.422674 INFO::Fitting model to feature number 21, Coprococcus
#> 2024-04-30 19:50:16.426826 INFO::Fitting model to feature number 22, DTU089
#> 2024-04-30 19:50:16.430927 INFO::Fitting model to feature number 23, Dialister
#> 2024-04-30 19:50:16.435011 INFO::Fitting model to feature number 24, Dorea
#> 2024-04-30 19:50:16.439111 INFO::Fitting model to feature number 25, Eggerthella
#> 2024-04-30 19:50:16.443247 INFO::Fitting model to feature number 26, Eisenbergiella
#> 2024-04-30 19:50:16.447282 INFO::Fitting model to feature number 27, Enterococcus
#> 2024-04-30 19:50:16.451322 INFO::Fitting model to feature number 28, Erysipelatoclostridium
#> 2024-04-30 19:50:16.455386 INFO::Fitting model to feature number 29, Erysipelotrichaceae.UCG.003
#> 2024-04-30 19:50:16.459468 INFO::Fitting model to feature number 30, Escherichia.Shigella
#> 2024-04-30 19:50:16.463606 INFO::Fitting model to feature number 31, Faecalibacterium
#> 2024-04-30 19:50:16.467686 INFO::Fitting model to feature number 32, Family.XIII.AD3011.group
#> 2024-04-30 19:50:16.471784 INFO::Fitting model to feature number 33, Flavonifractor
#> 2024-04-30 19:50:16.475857 INFO::Fitting model to feature number 34, Fusicatenibacter
#> 2024-04-30 19:50:16.487484 INFO::Fitting model to feature number 35, Fusobacterium
#> 2024-04-30 19:50:16.491797 INFO::Fitting model to feature number 36, Gemella
#> 2024-04-30 19:50:16.495917 INFO::Fitting model to feature number 37, Granulicatella
#> 2024-04-30 19:50:16.500037 INFO::Fitting model to feature number 38, Haemophilus
#> 2024-04-30 19:50:16.504143 INFO::Fitting model to feature number 39, Holdemania
#> 2024-04-30 19:50:16.508282 INFO::Fitting model to feature number 40, Hungatella
#> 2024-04-30 19:50:16.512392 INFO::Fitting model to feature number 41, Incertae.Sedis..EUPATH_0009256_Bacteria_Firmicutes_Clostridia_Oscillospirales_Ruminococcaceae_Incertae_Sedis.
#> 2024-04-30 19:50:16.516533 INFO::Fitting model to feature number 42, Intestinibacter
#> 2024-04-30 19:50:16.520617 INFO::Fitting model to feature number 43, Intestinimonas
#> 2024-04-30 19:50:16.524748 INFO::Fitting model to feature number 44, Klebsiella
#> 2024-04-30 19:50:16.528854 INFO::Fitting model to feature number 45, Lachnoclostridium
#> 2024-04-30 19:50:16.532944 INFO::Fitting model to feature number 46, Lachnospira
#> 2024-04-30 19:50:16.537066 INFO::Fitting model to feature number 47, Lachnospiraceae.FCS020.group
#> 2024-04-30 19:50:16.54113 INFO::Fitting model to feature number 48, Lachnospiraceae.ND3007.group
#> 2024-04-30 19:50:16.545205 INFO::Fitting model to feature number 49, Lachnospiraceae.NK4A136.group
#> 2024-04-30 19:50:16.549408 INFO::Fitting model to feature number 50, Lachnospiraceae.UCG.001
#> 2024-04-30 19:50:16.553497 INFO::Fitting model to feature number 51, Lachnospiraceae.UCG.004
#> 2024-04-30 19:50:16.557634 INFO::Fitting model to feature number 52, Lachnospiraceae.UCG.009
#> 2024-04-30 19:50:16.561731 INFO::Fitting model to feature number 53, Lachnospiraceae.UCG.010
#> 2024-04-30 19:50:16.565838 INFO::Fitting model to feature number 54, Lactobacillus
#> 2024-04-30 19:50:16.569911 INFO::Fitting model to feature number 55, Lactococcus
#> 2024-04-30 19:50:16.574012 INFO::Fitting model to feature number 56, Mannheimia
#> 2024-04-30 19:50:16.578088 INFO::Fitting model to feature number 57, Megasphaera
#> 2024-04-30 19:50:16.582188 INFO::Fitting model to feature number 58, Monoglobus
#> 2024-04-30 19:50:16.586308 INFO::Fitting model to feature number 59, Neisseria
#> 2024-04-30 19:50:16.59044 INFO::Fitting model to feature number 60, Odoribacter
#> 2024-04-30 19:50:16.594612 INFO::Fitting model to feature number 61, Oscillibacter
#> 2024-04-30 19:50:16.598837 INFO::Fitting model to feature number 62, Oscillospira
#> 2024-04-30 19:50:16.602924 INFO::Fitting model to feature number 63, Parabacteroides
#> 2024-04-30 19:50:16.607015 INFO::Fitting model to feature number 64, Parasutterella
#> 2024-04-30 19:50:16.611089 INFO::Fitting model to feature number 65, Prevotella
#> 2024-04-30 19:50:16.623001 INFO::Fitting model to feature number 66, Pseudomonas
#> 2024-04-30 19:50:16.627174 INFO::Fitting model to feature number 67, Romboutsia
#> 2024-04-30 19:50:16.631338 INFO::Fitting model to feature number 68, Roseburia
#> 2024-04-30 19:50:16.635478 INFO::Fitting model to feature number 69, Rothia
#> 2024-04-30 19:50:16.639635 INFO::Fitting model to feature number 70, Ruminococcus
#> 2024-04-30 19:50:16.643779 INFO::Fitting model to feature number 71, Sellimonas
#> 2024-04-30 19:50:16.647969 INFO::Fitting model to feature number 72, Solobacterium
#> 2024-04-30 19:50:16.652054 INFO::Fitting model to feature number 73, Staphylococcus
#> 2024-04-30 19:50:16.656152 INFO::Fitting model to feature number 74, Streptococcus
#> 2024-04-30 19:50:16.660285 INFO::Fitting model to feature number 75, Subdoligranulum
#> 2024-04-30 19:50:16.664358 INFO::Fitting model to feature number 76, Sutterella
#> 2024-04-30 19:50:16.668443 INFO::Fitting model to feature number 77, Terrisporobacter
#> 2024-04-30 19:50:16.67252 INFO::Fitting model to feature number 78, Turicibacter
#> 2024-04-30 19:50:16.67663 INFO::Fitting model to feature number 79, Tuzzerella
#> 2024-04-30 19:50:16.680714 INFO::Fitting model to feature number 80, Tyzzerella
#> 2024-04-30 19:50:16.684808 INFO::Fitting model to feature number 81, UBA1819
#> 2024-04-30 19:50:16.68892 INFO::Fitting model to feature number 82, UCG.002
#> 2024-04-30 19:50:16.692998 INFO::Fitting model to feature number 83, UCG.003
#> 2024-04-30 19:50:16.69709 INFO::Fitting model to feature number 84, UCG.005
#> 2024-04-30 19:50:16.701159 INFO::Fitting model to feature number 85, Veillonella
#> 2024-04-30 19:50:16.705248 INFO::Fitting model to feature number 86, X.Clostridium..innocuum.group
#> 2024-04-30 19:50:16.709336 INFO::Fitting model to feature number 87, X.Eubacterium..eligens.group
#> 2024-04-30 19:50:16.713444 INFO::Fitting model to feature number 88, X.Eubacterium..fissicatena.group
#> 2024-04-30 19:50:16.717537 INFO::Fitting model to feature number 89, X.Eubacterium..hallii.group
#> 2024-04-30 19:50:16.721626 INFO::Fitting model to feature number 90, X.Eubacterium..siraeum.group
#> 2024-04-30 19:50:16.725689 INFO::Fitting model to feature number 91, X.Eubacterium..ventriosum.group
#> 2024-04-30 19:50:16.7298 INFO::Fitting model to feature number 92, X.Eubacterium..xylanophilum.group
#> 2024-04-30 19:50:16.733902 INFO::Fitting model to feature number 93, X.Ruminococcus..gauvreauii.group
#> 2024-04-30 19:50:16.738032 INFO::Fitting model to feature number 94, X.Ruminococcus..gnavus.group
#> 2024-04-30 19:50:16.750366 INFO::Fitting model to feature number 95, X.Ruminococcus..torques.group
#> 2024-04-30 19:50:16.754814 INFO::Fitting model to feature number 96, unclassified.Chloroplast
#> 2024-04-30 19:50:16.759083 INFO::Fitting model to feature number 97, unclassified.Clostridia.UCG.014
#> 2024-04-30 19:50:16.763217 INFO::Fitting model to feature number 98, unclassified.Clostridia
#> 2024-04-30 19:50:16.767342 INFO::Fitting model to feature number 99, unclassified.Clostridia.vadinBB60.group
#> 2024-04-30 19:50:16.771517 INFO::Fitting model to feature number 100, unclassified.Gastranaerophilales
#> 2024-04-30 19:50:16.775677 INFO::Fitting model to feature number 101, unclassified.Lachnospiraceae
#> 2024-04-30 19:50:16.779801 INFO::Fitting model to feature number 102, unclassified.Mitochondria
#> 2024-04-30 19:50:16.78389 INFO::Fitting model to feature number 103, unclassified.Oscillospiraceae
#> 2024-04-30 19:50:16.78809 INFO::Fitting model to feature number 104, unclassified.Peptoniphilus
#> 2024-04-30 19:50:16.792167 INFO::Fitting model to feature number 105, unclassified.Rhodospirillales
#> 2024-04-30 19:50:16.796301 INFO::Fitting model to feature number 106, unclassified.Ruminococcaceae
#> 2024-04-30 19:50:16.800428 INFO::Fitting model to feature number 107, unclassified..Eubacterium..coprostanoligenes.group
#> 2024-04-30 19:50:16.821401 INFO::Counting total values for each feature
#> 2024-04-30 19:50:16.828032 INFO::Writing filtered data to file /tmp/RtmpfTYxYt/maaslin1ace1d0423d0/features/filtered_data.tsv
#> 2024-04-30 19:50:16.978624 INFO::Writing filtered, normalized data to file /tmp/RtmpfTYxYt/maaslin1ace1d0423d0/features/filtered_data_norm.tsv
#> 2024-04-30 19:50:17.132404 INFO::Writing filtered, normalized, transformed data to file /tmp/RtmpfTYxYt/maaslin1ace1d0423d0/features/filtered_data_norm_transformed.tsv
#> 2024-04-30 19:50:17.404628 INFO::Writing residuals to file /tmp/RtmpfTYxYt/maaslin1ace1d0423d0/fits/residuals.rds
#> 2024-04-30 19:50:17.476451 INFO::Writing fitted values to file /tmp/RtmpfTYxYt/maaslin1ace1d0423d0/fits/fitted.rds
#> 2024-04-30 19:50:17.502385 INFO::Writing all results to file (ordered by increasing q-values): /tmp/RtmpfTYxYt/maaslin1ace1d0423d0/all_results.tsv
#> 2024-04-30 19:50:17.504133 INFO::Writing the significant results (those which are less than or equal to the threshold of 0.250000 ) to file (ordered by increasing q-values): /tmp/RtmpfTYxYt/maaslin1ace1d0423d0/significant_results.tsv